### 2025-02-12 #7776 @editorialbot
[PRE REVIEW]: pyphot: A tool for computing photometry from spectra
**Submitting author:** @mfouesneau (<a href="http://orcid.org/0000-0001-9256-5516">Morgan Fouesneau</a>)
**Repository:** https://github.com/mfouesneau/pyphot
**Branch with paper.md** (empty if default branch):
**Version:** v1.6.0
