### 2025-02-12 #7776 @warrickball
While we wait, could you also comment on the selection of authors for the paper? Specifically, why is Ariane Lancon an author when other apparent contributors are not? I don't doubt there's a reason but we'd like to have it recorded ahead of a review. Thanks!
