### 2025-02-18 #7776 @mfouesneau
@warrickball, thanks for the updates.

Ariane Lancon has significantly contributed to the original project Pyphot spawned from, especially on the LIS library.
