+++ _sources/photometry.rst
@@ -1,0 +1,283 @@
+Details on predicting photometry
+================================
+
+It is sometimes not obvious that there are important differences between
+photometric systems. But even less known the difference between detector
+count types (energy or photons) which requires also special care.
+
+In this section, we review the important details for computing the luminosity
+and the magnitude of a star through a photometric passband filter. We do not
+discuss calibration, which in principle is covered by instrument documentations.
+
+If we consider a filter throughput (a.k.a, transmission curve, or response
+function) defined in wavelength by the dimensionless function :math:`T(\lambda)`, 
+this function tells you what fraction of the arriving photons at wavelength
+:math:`\lambda` actually get through the instrument.  Therefore, the total number of
+photons, per unit time per unit area, received in this filter is
+
+.. math::
+
+        \begin{equation}
+        N_{tot} = \frac{1}{hc} \int_\lambda f_\lambda\,\lambda\,T(\lambda)\,d\lambda,
+        \end{equation}
+
+where :math:`f_\lambda` is the wavelength dependent flux density of an object
+given in energy per unit time per unit area per unit wavelength.
+
+Consequently, interpreting :math:`\lambda T(\lambda)` as a distribution leads to
+the statistical mean of the flux density, :math:`\overline{f_\lambda}` 
+
+.. math::
+
+        \begin{equation}
+        \overline{f_\lambda}(T) = \frac{\int_\lambda \lambda f_\lambda T(\lambda) d\lambda}{\int_\lambda \lambda T(\lambda) d\lambda}.
+        \end{equation}
+
+Note that this is not the mean flux density because of the :math:`\lambda` factor in
+the integrals, it is the mean photon rate density in this filter commonly
+expressed in stellar physics literature as :math:`erg/s/cm^2/\unicode{x212B}` or 
+:math:`W/m^2/\unicode{x212B}`.
+
+.. code-block:: python
+
+        # computing the flux of a spectrum
+        flux = lib['hst_wfc3_f110w'].get_flux(lamb, spec)
+        # lamb may have units, otherwise assuming consistent definitions.
+
+        # computing the flux of many spectra
+        fluxes = lib['hst_wfc3_f110w'].get_flux(lamb, spectra, axis=1)
+
+
+Finally, at least for instruments using CCD or CCD-like cameras, i.e., counting
+photons, we obtain the usual definition of magnitude 
+
+.. math::
+
+        \begin{equation}
+        mag_\lambda(T) = -2.5\,\log_{10}\left(\overline{f_\lambda}\right) - ZP\left(\overline{f_\lambda}\right),
+        \end{equation}
+
+where :math:`ZP(\overline{f_\lambda})` gives the pass-band reference value
+(zeropoint) for a given photometric/magnitude system.
+
+However, the zeropoints themselves depend on the adopted photometric system used
+to report the measurements. They may vary fundamentally from one to another.
+Below we briefly describe the main systems used in large surveys.
+
+
+
+Vega magnitude system
+~~~~~~~~~~~~~~~~~~~~~
+
+This system is defined such that the star Alpha Lyr (Vega) has magnitude 0 in
[…]
