+++ _sources/pyphot.rst
@@ -1,0 +1,110 @@
+pyphot package
+==============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyphot.astropy
+   pyphot.ezunits
+
+Submodules
+----------
+
+pyphot.config module
+--------------------
+
+.. automodule:: pyphot.config
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.helpers module
+---------------------
+
+.. automodule:: pyphot.helpers
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.licks module
+-------------------
+
+.. automodule:: pyphot.licks
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.pbar module
+------------------
+
+.. automodule:: pyphot.pbar
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.phot module
+------------------
+
+.. automodule:: pyphot.phot
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.sandbox module
+---------------------
+
+.. automodule:: pyphot.sandbox
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.simpletable module
+-------------------------
+
+.. automodule:: pyphot.simpletable
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.sun module
+-----------------
+
+.. automodule:: pyphot.sun
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.svo module
+-----------------
+
+.. automodule:: pyphot.svo
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.vega module
+------------------
+
+.. automodule:: pyphot.vega
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.version module
+---------------------
+
+.. automodule:: pyphot.version
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: pyphot
+   :members:
+   :show-inheritance:
+   :undoc-members:
