### 2025-05-01 858b5b3
Update index.rst
Updated the outputs in the example exploration of one IRAC transmission curve.
