+++ doc/index.rst
@@ -119,6 +119,8 @@ Quick Start
 
 Additional examples can be found on the :doc:`more_examples` page.
 
+If one wishes to find out details about one of the available transmission curves:
+
 .. code-block:: python
 
         import pyphot
@@ -133,7 +135,6 @@ Additional examples can be found on the :doc:`more_examples` page.
 
 .. code-block:: none
 
-        Library contains:  196  filters
         Filter object information:
             name:                 SPITZER_IRAC_45
             detector type:        photon
@@ -141,24 +142,30 @@ Additional examples can be found on the :doc:`more_examples` page.
             central wavelength:   45110.141614 angstrom
             pivot wavelength:     45020.219955 angstrom
             effective wavelength: 44425.747085 angstrom
+            photon wavelength:    44603.204646 angstrom
+            minimum wavelength:   39250.000000 angstrom
+            maximum wavelength:   50640.000000 angstrom
             norm:                 4664.680820
+            effective width:      8714.143135 angstrom
+            fullwidth half-max:   10110.000000
             definition contains 417 points
 
             Zeropoints
-                Vega: 28.933084 mag,
-                      2.671569250882836e-12 erg / angstrom * centimeter ** 2 * second,
-                      175.8794962126167 Jy
+                Vega: 28.933083 mag,
+                      2.6715713304827696e-12 erg / angstrom * centimeter ** 2 * second,
+                      180.61811118349118 Jy
+                      3.246733893355585 photon / angstrom * centimeter ** 2 * second
                   AB: 25.674986 mag,
                       5.370385702161592e-11 erg / angstrom * centimeter ** 2 * second,
-                      3535.5277855945205 Jy
+                      3630.780547701007 Jy
                   ST: 21.100000 mag,
                       3.6307805477010028e-09 erg / angstrom * centimeter ** 2 * second,
-                      239027.9995089771 Jy 
+                      245467.79536259372 Jy
 
         [...]
 
 Suppose one has a calibrated spectrum and wants to compute the vega magnitude
-throug the HST WFC3 F110W passband, 
+through the HST/WFC3 F110W passband: 
 
 .. code-block:: python
 
