### 2025-05-01 443bfab
Update index.rst
Small text-edits for the two last example-boxes of the Quick-start section.
