+++ doc/index.rst
@@ -164,25 +164,25 @@ If one wishes to find out details about one of the available transmission curves
 
         [...]
 
-Suppose one has a calibrated spectrum and wants to compute the vega magnitude
+Suppose one has a calibrated spectrum and wants to compute the corresponding vega magnitude
 through the HST/WFC3 F110W passband: 
 
 .. code-block:: python
 
-        # convert to magnitudes
         import numpy as np
         f = lib['hst_wfc3_f110w']
         # compute the integrated flux through the filter f
-        # note that it work on many spectra at once
+        # (note that 'spectra' can be a 2D array holding many spectra)
         fluxes = f.get_flux(lamb, spectra, axis=1)
-        # convert to vega magnitudes
+        # convert fluxes to vega magnitudes
         mags = -2.5 * np.log10(fluxes) - f.Vega_zero_mag
         # or similarly
         mags = -2.5 * np.log10(fluxes / f.Vega_zero_flux)
 
 
-If one wants to use a given transmission curve as filter, defined by `lamb_T` and
-`T`, one would use the :class:`pyphot.phot.Filter` directly as 
+To use one's own transmission curve, defined by `lamb_T` and
+`T`, instead of one of the pre-defined ones, one would use 
+the :class:`pyphot.phot.Filter` directly: 
 
 .. code-block:: python
 
@@ -245,7 +245,7 @@ Author: Morgan Fouesneau (`@mfouesneau <https://github.com/mfouesneau>`_)
 
 Direct contributions to the code base:
 
-* Ariane Lancon (`@lancon <https://github.com/lancon>`_)
+* Ariane Lançon (`@lancon <https://github.com/lancon>`_)
 * Tim Morton (`@timothydmorton <https://github.com/timothydmorton>`_)
 
 Related projects
