### 2025-06-11 47177d0
Update photometry.rst
Updates to clarify that "filter" can be more than just the filter (and it is up to the user); plus some formulation changes.
