+++ doc/photometry.rst
@@ -2,18 +2,28 @@ Details on predicting photometry
 ================================
 
 It is sometimes not obvious that there are important differences between
-photometric systems. But even less known the difference between detector
+photometric systems. But even less known is the difference between detector
 count types (energy or photons) which requires also special care.
 
 In this section, we review the important details for computing the luminosity
-and the magnitude of a star through a photometric passband filter. We do not
+and the magnitude of a star through a photometric passband. We do not
 discuss calibration, which in principle is covered by instrument documentations.
 
+Synthetic photometry uses transmission curves that typically include a variety of
+wavelength-dependent components (a filter transmission, the response of the optical
+elements of the telescope and instrument, the detector sensitivity, sometimes a telluric 
+absorption component, ..). Transmission curves are usually named after the associated filter, 
+and it is up to the user to verify that the other components are also included. In this
+documentation, we use the terms filter throughput, transmission curve, response function,
+or simply filter, interchangeably, to refer to a transmission curve, generally leaving the 
+verification of the nature of these transmission curves to the user. 
+
+Most modern detectors count photons, rather than cumulating the energies of these photons. 
 If we consider a filter throughput (a.k.a, transmission curve, or response
 function) defined in wavelength by the dimensionless function :math:`T(\lambda)`, 
 this function tells you what fraction of the arriving photons at wavelength
-:math:`\lambda` actually get through the instrument.  Therefore, the total number of
-photons, per unit time per unit area, received in this filter is
+:math:`\lambda` actually get detected (on average).  Therefore, the total number of
+photons, per unit time per unit collecting area, expected to be detected in this filter is
 
 .. math::
 
@@ -22,7 +32,7 @@ photons, per unit time per unit area, received in this filter is
         \end{equation}
 
 where :math:`f_\lambda` is the wavelength dependent flux density of an object
-given in energy per unit time per unit area per unit wavelength.
+given in energy per unit time, per unit collecting area, and per unit wavelength.
 
 Consequently, interpreting :math:`\lambda T(\lambda)` as a distribution leads to
 the statistical mean of the flux density, :math:`\overline{f_\lambda}` 
@@ -33,9 +43,9 @@ the statistical mean of the flux density, :math:`\overline{f_\lambda}`
         \overline{f_\lambda}(T) = \frac{\int_\lambda \lambda f_\lambda T(\lambda) d\lambda}{\int_\lambda \lambda T(\lambda) d\lambda}.
         \end{equation}
 
-Note that this is not the mean flux density because of the :math:`\lambda` factor in
-the integrals, it is the mean photon rate density in this filter commonly
-expressed in stellar physics literature as :math:`erg/s/cm^2/\unicode{x212B}` or 
+Note that although this is formally a weighted mean of a flux density, with weights proportional to :math:`\lambda T(\lambda)` (the denominator ensuring that the sum of the weights is 1), it actually
+measures the mean photon rate density in this filter. The result is commonly
+expressed in the same units as :math:`f_\lambda`, i.e. :math:`erg/s/cm^2/\unicode{x212B}` or 
 :math:`W/m^2/\unicode{x212B}`.
 
 .. code-block:: python
[…]
