### 2025-07-01 c31e8a3
Adding flexible flavors of Vega
