+++ pyphot/astropy/vega.py
@@ -1,27 +1,40 @@
-""" Handle vega spec/mags/fluxes manipulations
+"""Handle vega spec/mags/fluxes manipulations
 
 Works with both ascii and hd5 files for back-compatibility
 
 Vega.wavelength and Vega.flux have now units!
 """
+
 from __future__ import print_function
+
+import warnings
 from functools import wraps
+
 import numpy
-from .config import libsdir
-from ..simpletable import SimpleTable
 from astropy.units import Unit
 
+from ..simpletable import SimpleTable
+from .config import libsdir
 
-__all__ = ['Vega', 'from_Vegamag_to_Flux', 'from_Vegamag_to_Flux_SN_errors']
+__all__ = ["Vega", "from_Vegamag_to_Flux", "from_Vegamag_to_Flux_SN_errors"]
 
-_default_vega = "{0}/vega.hd5".format(libsdir)
+_default_vega = {
+    "mod_002": "{0}/alpha_lyr_mod_002.fits".format(libsdir),
+    "mod_003": "{0}/alpha_lyr_mod_003.fits".format(libsdir),
+    "mod_004": "{0}/alpha_lyr_mod_004.fits".format(libsdir),
+    "stis_011": "{0}/alpha_lyr_stis_011.fits".format(libsdir),
+    "legacy": "{0}/vega.hd5".format(libsdir),
+}
 
 
 class Vega(object):
     """
     Class that handles vega spectrum and references.  This class know where to
-    find the Vega synthetic spectrum (Bohlin 2007) in order to compute fluxes
+    find the Vega synthetic spectrum in order to compute fluxes
     and magnitudes in given filters
+
+    Default Vega spectrum is Bohlin 2007, alpha_Lyr_mod_003.fits
+    from the HST CDBS database, which is a synthetic spectrum of Vega
 
     Attributes
     ----------
@@ -31,10 +44,8 @@ class Vega(object):
         data table
     units: tuple
         detected units from file header
-    wavelength: array
-        wavelength (with units when found)
-    flux: array
-        flux(wavelength) values (with units when provided)
+    flavor: str, (default theoretical)
+        ["mod_002", "mod_003", "mod_004"] or ["stis_011"]
 
     An instance can be used as a context manager as:
 
@@ -45,9 +56,24 @@ class Vega(object):
         print vega_f, vega_mag, flamb
     """
 
-    def __init__(self, source=_default_vega):
-        """ Constructor """
-        self.source = source
+    def __init__(self, source=None, flavor="legacy"):
+        """Constructor"""
+        self.data = None
+        self.units = None
+        self._set_source_flavor(source, flavor)
+
+    def _set_source_flavor(self, source=None, flavor="mod_003"):
+        """Set the source and flavor of the Vega spectrum"""
+        if source is not None:
+            self.source = source
+        else:
+            if flavor.lower() not in _default_vega:
+                raise ValueError(
+                    "Unknown Vega flavor: {0}. Available flavors {1}".format(
+                        flavor, _default_vega.keys()
+                    )
+                )
+            self.source = _default_vega[flavor]
         self.data = None
         self.units = None
 
@@ -56,32 +82,44 @@ class Vega(object):
             return
         fname = self.source
         # get extension
-        ext = fname.split('.')[-1]
-        if ( ext.lower() in ('hd5', 'hdf', 'hdf5') ):
-            self.data = SimpleTable(fname, '/spectrum', silent=True)
+        ext = fname.split(".")[-1]
+        if ext.lower() in ("hd5", "hdf", "hdf5"):
+            self.data = SimpleTable(fname, "/spectrum", silent=True)
         else:
[…]
