### 2025-07-01 0fdcd36
Capturing astropy warning on fits
