### 2025-07-01 9ddb685
doc on vega spectra
