### 2025-07-01 8b2e26a
minor vega documentation
