+++ doc/vega.rst
@@ -63,7 +63,9 @@ Impact of Vega Flavors on zeropoints provided by Pyphot:
 * Changes in temperature will introduce a wavelength dependent shift in the vega zero points.
 * Changes in logg, metallicity, turbulence will introduce non trivial variations in the vega zero points.
 
-By default, Pyphot uses the `alpha_stis_003` flavor as the Vega standard, which may be updated to `alpha_stis_011` in the future.
+.. note:: 
+
+    By default, Pyphot uses the `alpha_stis_003` flavor as the Vega standard, which may be updated to `alpha_stis_011` in the future.
 
 
 TODO: 
