### 2025-07-01 8fd722a
added quick example
