### 2025-07-01 2aa69a7
Update vega.rst
minor
