### 2025-07-01 3f06f79
Update CITATION.cff version
