+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Morgan"
   orcid: "https://orcid.org/0000-0001-9256-5516"
 title: "pyphot"
-version: 1.4.3
-date-released: 2022-03-29
+version: 1.6.0
+date-released: 2025-01-21
 url: "https://github.com/mfouesneau/pyphot"
