### 2025-07-01 92a9436
Update vega.rst todolist
