### 2025-07-02 3ab2a8d
Adding vega documentation plots
