+++ doc/vega.rst
@@ -47,6 +47,13 @@ while addressing some of Vega's limitations (`Fukugita et al. 1996
 Vega Flavors in Pyphot
 ----------------------
 
+.. important::
+
+    Impact of Vega Flavors on zeropoints provided by Pyphot:
+
+    * Changes in temperature will introduce a wavelength dependent shift in the vega zero points.
+    * Changes in logg, metallicity, turbulence will introduce non trivial variations in the vega zero points.
+
 Since version 1.7.0, Pyphot includes a set of Vega flavors one can use transparently as photometric standards. We summarize the available flavors below:
 
 * `alpha_lyr_004` from `Bohlin, Colina, & Finley (1995) <https://ui.adsabs.harvard.edu/abs/1995AJ....110.1316B>`_. It corrsponds to pure hydrogen white dwarf models are the absolute flux standards.
@@ -58,30 +65,138 @@ Since version 1.7.0, Pyphot includes a set of Vega flavors one can use transpare
 * `alpha_mod_004` from `Bohlin, Hubeny, Rauch (2020) <https://ui.adsabs.harvard.edu/abs/2020AJ....160...21B>`_ is similar to `alpha_mod_003` but with a Flux(5556A)=3.47e-9. In the details, it also includes additional lines.
 * `alpha_stis_011` is a special model from Bohlin which is a composite flux of a special Kurucz 9550K model from 900-1152A (Kurucz 2003), IUE data from 1152-1675A, STIS CCD fluxes from 1675-10200A (Bohlin & Gilliland 2004a), and the 9550K model longward of 10200A. It differs significantly in the uv-optical range by accounting for the new TMAP AND TLUSTY WD NLTE models (`Bohlin, Hubeny, Rauch (2020) <https://ui.adsabs.harvard.edu/abs/2020AJ....160...21B>`_) but corresponds to `alpha_mod_004` above 1 micron.
 
-Impact of Vega Flavors on zeropoints provided by Pyphot:
-
-* Changes in temperature will introduce a wavelength dependent shift in the vega zero points.
-* Changes in logg, metallicity, turbulence will introduce non trivial variations in the vega zero points.
-
-.. code-block:: python
-
-    import pyphot
-    from pyphot import vega
-
-    vega_s011 = Vega(flavor="stis_011") 
-    vega_m004 = Vega(flavor="mod_004")
-    vega_default = Vega()  # uses the default flavor, which is stis_003
-
-
 .. note:: 
 
     By default, Pyphot uses the `alpha_stis_003` flavor as the Vega standard, which may be updated to `alpha_stis_011` in the future.
 
+The figure below shows the differences between the various Vega flavors
+available in Pyphot. The top panel shows the absolute fluxes, while the bottom
+panel shows the relative differences with respect to the `alpha_stis_003` (the
+legacy Pyphot reference flavor).  flavor.
 
-TODO
-----
+.. figure:: vega_flavors_comparison.png
+   :align: center
+   :scale: 70 %
 
-* make a plots that shows major differences and their scale: ~percent and below for minor updates, up to 10% in the UV for  `alpha_mod_004` vs `alpha_stis_011`.
+   **Figure:**  Variations of Vega Flavors relative to Pyphot legacy reference flavor (`alpha_stis_003`).
[…]
