### 2025-07-02 a7a0183
adding reset and set vega
