+++ doc/vega.rst
@@ -44,30 +44,40 @@ the AB magnitude system providing an alternative that maintains compatibility
 while addressing some of Vega's limitations (`Fukugita et al. 1996
 <https://ui.adsabs.harvard.edu/abs/1996AJ....111.1748F>`_).
 
-Vega Flavors in Pyphot
-----------------------
+Vega Flavors
+------------
+
 
 .. important::
+
+    Since version 1.6.1, Pyphot includes a set of Vega flavors one can use transparently as photometric standards.
+    
+    * Pyphot provides `alpha_stis_003` (aliased to `legacy`) as the default Vega flavor (and only one before version 1.6.1). This flavor is also the reference for the values provided in the documentation.
+    * (since version 1.6.1) Pyphot provides also `alpha_lyr_mod_003`, `alpha_lyr_mod_004`, and `alpha_stis_011` as alternative Vega flavors.
+    * in a future release, Pyphot will switch the default Vega flavor to `alpha_stis_011` which is a more recent model that incorporates updated NLTE models and provides a more accurate representation of Vega's spectral energy distribution, particularly in the ultraviolet and optical ranges.
 
     Impact of Vega Flavors on zeropoints provided by Pyphot:
 
     * Changes in temperature will introduce a wavelength dependent shift in the vega zero points.
     * Changes in logg, metallicity, turbulence will introduce non trivial variations in the vega zero points.
 
-Since version 1.7.0, Pyphot includes a set of Vega flavors one can use transparently as photometric standards. We summarize the available flavors below:
+
+
+`CALSPEC
+<https://www.stsci.edu/hst/instrumentation/reference-data-for-calibration-and-tools/astronomical-catalogs/calspec>`_
+provides a collection of composite stellar spectra that are flux standards for
+the calibration of systems such as HST. In particular, it compiles many flavors
+of Vega of which we summarize the most relevant ones below: 
 
 * `alpha_lyr_004` from `Bohlin, Colina, & Finley (1995) <https://ui.adsabs.harvard.edu/abs/1995AJ....110.1316B>`_. It corrsponds to pure hydrogen white dwarf models are the absolute flux standards.
 * `alpha_lyr_005` from Colina, Bohlin & Castelli (1996), Instrument Science Report (`OSG-CAL-96-01 <https://www.stsci.edu/instruments/observatory/PDF/scs8.rev.pdf>`_). It corresponds to the combination of ultraviolet covered by average IUE spectrum; Optical up to 1.05 microns covered by Hayes (1985) average spectrum (IAU Symp 111, p 225); near-infrared covered by ATLAS12 based model computed by Dr. Castelli, rebinned to 25A, and normalized to Hayes (1985) Johnson V flux. V filter as in Buser & Kurucz 1979, AA 70, 555. (All wavelengths are vacuum.)
-* `alpha_stis_003` (legacy) from `Bohlin (2007) <https://ui.adsabs.harvard.edu/abs/2007ASPC..364..315B>`_ a special model from Kurucz 9400K Vega spectrum T/g=9400/3.9 [M/H]=-0.5 fit against STIS data with microturbulence of 0. km/s, and where V = 0.023 mag.
[…]
