+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyphot"
-version = "1.6.0"
+version = "1.7.0"
 description = "A tool for computing photometry from spectra"
 readme = "README.rst"
 requires-python = ">=3.9"
