+++ _sources/pyphot.astropy.rst
@@ -1,0 +1,45 @@
+pyphot.astropy package
+======================
+
+Submodules
+----------
+
+pyphot.astropy.config module
+----------------------------
+
+.. automodule:: pyphot.astropy.config
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.astropy.sandbox module
+-----------------------------
+
+.. automodule:: pyphot.astropy.sandbox
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.astropy.sun module
+-------------------------
+
+.. automodule:: pyphot.astropy.sun
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.astropy.vega module
+--------------------------
+
+.. automodule:: pyphot.astropy.vega
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: pyphot.astropy
+   :members:
+   :show-inheritance:
+   :undoc-members:
