+++ _sources/index.rst
@@ -1,0 +1,322 @@
+.. pyphot documentation master file, created by
+   sphinx-quickstart on Wed Oct  5 11:25:47 2016.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+PYPHOT -- A tool for computing photometry from spectra
+======================================================
+
+This is a set of tools to compute synthetic photometry in a simple way, suitable for
+integration in larger projects.
+
+The inputs are response functions for the desired photometric passbands, and stellar spectra. The modules are flexible to handle units in the wavelength definition through 
+a simplified version of `pint` (link)
+
+Filters are represented individually by a `Filter` object. Collections of
+filters are handled with a `Library`. We provide an internal library that
+contains a signitificant amount of common filters. Note that we use the word "filter" 
+in a generic sense: in many cases it corresponds in fact to a total throughput, that includes
+the actual filter transmission, the transmission of the optics, the wavelength-dependent efficiency
+of the detector, and (for ground-based photometric systems) the transmission of the 
+atmosphere.
+
+Each filter is minimally defined by a `wavelength`, a `throughput`, and a detector
+type that may be either `energy` or `photon` (default). Many properties such as central 
+or pivot wavelengths are computed internally. 
+
+Units are provided for the wavelengths of the filters, zero points in
+multiple units are also accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The
+default detector is assumed to be a photon-counting device, but energy-sensitive
+detectors are also handled for the computations.
+
+.. note::
+
+        All provided filters have defined units and detector type
+        that are used transparently throughout the package.
+
+.. image:: https://mybinder.org/badge.svg 
+  :target: https://mybinder.org/v2/gh/mfouesneau/pyphot/master?filepath=examples%2FQuickStart.ipynb
+ 
+.. image:: https://img.shields.io/badge/render%20on-nbviewer-orange.svg
+  :target: https://nbviewer.jupyter.org/github/mfouesneau/pyphot/tree/master/examples/
+  
+
+Package main content
+~~~~~~~~~~~~~~~~~~~~
+
+This package is mainly organized around two backend to handle quantities with units: a lightweight version of
+`Pint`_ (:class:`pyphot.ezunits`) and the `Astropy Units <https://docs.astropy.org/en/stable/units/index.html>`_. 
+As these do no provide the same functionalities, the package is designed to be able to switch between the two by importing either from :class:`pyphot.sandbox` or :class:`pyphot.astropy`.
+
+.. _Pint: https://pint.readthedocs.io/en/stable/
+
+.. tip::
+
+        You can transparently switch between pint and astropy units with aliasing the import line:
+
+        .. code:: python
+
+                from pyphot import astropy as pyphot
+                # or 
+                from pyphot import sandbox as pyphot
+
+
+This package is mostly organized around 2 main classes (from either root module above):
+
+* :class:`~pyphot.sandbox.UnitFilter` that handles filter definitions and manipulations.
+ 
+* :class:`~pyphot.sandbox.UnitLibrary` that handles a collection of filters in a few formats. 
+
+
+Additionally, and for convenience 
+
+* the library class is derived into an ascii file reader
+  :class:`~pyphot.phot.Ascii_Library`, and a single hdf file reader
+  :class:`~pyphot.phot.HDF_Library`. 
+
[…]

[…]
