### 2025-08-06 #7776 @warrickball
Hi @mfouesneau, sorry for the very long wait. Your paper is now in essence at the front of the queue, so it depends on an editor becoming available. We haven't expanded the editor pool just yet and a few are on vacation (and I will be myself for part of August) but hopefully we'll be able to start by the end of August.

Again, I apologise that it's now been about six months since you submitted. 😞
