### 2025-08-13 #7776 @WillJRoper
I'd be game but won't have free time to start for at least a week, maybe 2. If that's ok, you can sign me up. Equally happy if @christopherlovell or @aswinpvijayan want to and can get there first.
