### 2025-08-14 #8814 @editorialbot
[REVIEW]: pyphot: A tool for computing photometry from spectra
**Submitting author:** @mfouesneau (<a href="http://orcid.org/0000-0001-9256-5516">Morgan Fouesneau</a>)
**Repository:** https://github.com/mfouesneau/pyphot
**Branch with paper.md** (empty if default branch):
**Version:** v1.6.0
