### 2025-10-07 #8814 @WillJRoper
> [@JBorrow](https://github.com/JBorrow) [@WillJRoper](https://github.com/WillJRoper) [@christopherlovell](https://github.com/christopherlovell), sorry to ask, but is there any issue here? The paper was submitted in February

Hi @mfouesneau, sorry for not getting around to this yet. It's job application season, and there's been some other unforeseen stuff going on. I understand your frustration. We are also waiting for a review. I'll get round to this as soon as I can.
