### 2025-10-27 mfouesneau/pyphot#87 @christopherlovell
Naming and source of data in the source code directory
The data for the included filters is explicitly stored within the source directory, which is fine, but the package would benefit from better naming, perhaps a “data” directory, rather than “lib”, which is typically reserved for libraries in unix like systems. It is also unclear where this data has come from, is this a local copy of the SVO database or are these filters collected individually from other sources? If the latter, what are these sources, and are these referenced correctly in the source / documentation?

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8814#issuecomment-3451697350
