### 2025-10-27 mfouesneau/pyphot#88 @christopherlovell
Documentation issues
- The documentation does not clearly state what problem the software is designed to solve, nor the target audience.
- The documentation does not explicitly state what versions of python are supported, nor what the dependencies are.
- The Zenodo badge link on the README is broken
- There are no clear contribution guidelines, nor directions on where or how to raise issues with the code, nor guidance on seeking support. These should at minimum be provided within the README, with guidance on the code style required, format of PRs (is there a checklist contributors should follow?), and contact details. The latter could just be directions to raise issues if users have support requests, but this should be added explicitly. Ideally all of this information should also be provided within the documentation.
- Related to the above, it is not clear what the docstring style actually is. It seems to be Numpydoc, but it does not appear to be abided by throughout the source code. Please update all docstrings to a consistent format, and state the adopted style in the contribution guidelines.
- The package appears to be published on pypi, but install instructions in the documentation are only for the latest source code version. Install instructions for the latest pypi version through pip should be provided.
- There is a statement ‘maybe astropy’ in the example usage section. Since astropy is already implemented this should be fixed.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8814#issuecomment-3451697350
