### 2025-10-27 mfouesneau/pyphot#89 @christopherlovell
Test suite issues
There is only one unit test script, test_quick.py, and there are no code coverage statistics provided. The unit test suite would benefit from a number of significant improvements before I can recommend the package for publication in JOSS. This includes:
- Separate tests scripts for different sections of the code, with better demarcation of tested functionality.
- A code coverage badge on the front page, and some demonstration that a significant proportion of the code is covered.
- Some evidence of integration tests of complex functionality.
- Each unit test should have a docstring describing the purpose and target of the test.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8814#issuecomment-3451697350
