+++ pyphot/pbar.py
@@ -11,19 +11,16 @@ meter, the running time, an eta, and the rate of the iterations.
 An example is::
     description    [----------] k/n  10% [time: 00:00:00, eta: 00:00:00, 2.7 iters/sec]
 """
+
 import time as _time
 import sys
 import signal
 from array import array
-
-try:
-    from fcntl import ioctl
-    import termios
-except ImportError:
-    pass
+from fcntl import ioctl
+import termios
 
 
-__all__ = ['Pbar']
+__all__ = ["Pbar"]
 
 
 class Pbar(object):
@@ -66,12 +63,26 @@ class Pbar(object):
     units: str, optional (default: 'iters')
         unit of the iteration
     """
-    def __init__(self, maxval=None, desc=None, time=True, eta=True, rate=True, length=None,
-                 file=None, keep=True, mininterval=0.5, miniters=1, units='iters', **kwargs):
+
+    def __init__(
+        self,
+        maxval=None,
+        desc=None,
+        time=True,
+        eta=True,
+        rate=True,
+        length=None,
+        file=None,
+        keep=True,
+        mininterval=0.5,
+        miniters=1,
+        units="iters",
+        **kwargs,
+    ):
         self.time = time
         self.eta = eta
         self.rate = rate
-        self.desc = desc or ''
+        self.desc = desc or ""
         self.units = units
         self.file = file or sys.stdout
         self._last_print_len = 0
@@ -83,14 +94,14 @@ class Pbar(object):
         if length is not None:
             self.length = length
             self._auto_width = False
-        #backward compatibility
+        # backward compatibility
         self._start_t = _time.time()
         self._maxval = maxval
-        if 'txt' in kwargs:
-            self.desc = kwargs['txt']
+        if "txt" in kwargs:
+            self.desc = kwargs["txt"]
 
     def _buffer_width(self):
-        """ returns the width of the buffer when available """
+        """returns the width of the buffer when available"""
         try:
             self.handle_resize(None, None)
             signal.signal(signal.SIGWINCH, self.handle_resize)
@@ -102,12 +113,13 @@ class Pbar(object):
         return self.term_width
 
     def handle_resize(self, signum, frame):
-        h, w = array('h', ioctl(self.file, termios.TIOCGWINSZ, '\0' * 8))[:2]
+        ioinfo = ioctl(self.file, termios.TIOCGWINSZ, "\0" * 8)  # type: ignore
+        h, w = array("h", ioinfo)[:2]
         self.term_width = w
 
     @staticmethod
     def format_interval(t):
-        """ make a human readable time interval decomposed into days, hours,
+        """make a human readable time interval decomposed into days, hours,
         minutes and seconds
 
         Parameters
@@ -125,11 +137,11 @@ class Pbar(object):
         h, m = divmod(mins, 60)
         d, h = divmod(h, 24)
 
-        txt = '{m:02d}:{s:02d}'
+        txt = "{m:02d}:{s:02d}"
         if h:
-            txt = '{h:02d}:' + txt
+            txt = "{h:02d}:" + txt
         if d:
-            txt = '{d:d}d ' + txt
+            txt = "{d:d}d " + txt
         return txt.format(d=d, h=h, m=m, s=s)
 
     def build_str_meter(self, n, total, elapsed):
@@ -157,36 +169,36 @@ class Pbar(object):
         if n > total:
             total = None
 
-        vals = {'n': n}
-        vals['elapsed'] = self.format_interval(elapsed)
-        vals['rate'] = '{0:5.2f}'.format((n / elapsed)) if elapsed else '?'
-        vals['units'] = self.units
+        vals = {"n": n}
[…]
