+++ pyphot/__init__.py
@@ -1,4 +1,7 @@
-# from .phot import Library, Ascii_Library, HDF_Library, Filter
+from typing import Optional
+
+from pyphot.future.unit_adapters.units_adapter import UnitsAdapter
+
 from .licks import LickLibrary, LickIndex, reduce_resolution
 
 from .sandbox import (
@@ -12,25 +15,12 @@ from .sandbox import (
 
 from .sandbox import get_library
 from .ezunits import unit
-# from .helpers import (
-#     STmag_from_flux,
-#     STmag_to_flux,
-#     extractPhotometry,
-#     extractSEDs,
-#     fluxErrTomag,
-#     fluxToMag,
-#     magErrToFlux,
-#     magToFlux,
-# )
+
 from .vega import Vega
 from .sun import Sun
 from .version import __VERSION__
 
 __all__ = [
-    # Library,
-    # Ascii_Library,
-    # HDF_Library,
-    # Filter,
     LickLibrary,
     LickIndex,
     reduce_resolution,
@@ -42,14 +32,6 @@ __all__ = [
     UnitLickLibrary,
     get_library,
     unit,
-#     STmag_from_flux,
-#     STmag_to_flux,
-#     extractPhotometry,
-#     extractSEDs,
-#     fluxErrTomag,
-#     fluxToMag,
-#     magErrToFlux,
-#     magToFlux,
     Vega,
     Sun,
     __VERSION__,
