### 2025-10-31 e5c150a
add future adapters for unit packages
