+++ pyphot/future/unit_adapters/ezunits/pint.py
@@ -1,0 +1,1352 @@
+# -*- coding: utf-8 -*-
+"""
+pint
+~~~~
+
+Pint is Python module/package to define, operate and manipulate
+**physical quantities**: the product of a numerical value and a
+unit of measurement. It allows arithmetic operations between them
+and conversions from and to different units.
+(https://github.com/hgrecco/pint)
+
+:copyright: 2012 by Hernan E. Grecco.
+:license: BSD, see LICENSE for more details.
+
+"""
+
+from __future__ import division, print_function, absolute_import
+
+__version__ = "0.1"
+
+import sys
+import copy
+import math
+import logging
+import operator
+import functools
+import itertools
+
+try:
+    # python >= 3.10
+    from collections.abc import Iterable
+except ImportError:
+    from collections import Iterable
+
+# from pkg_resources import resource_filename
+from importlib import resources
+
+from io import BytesIO
+from numbers import Number
+import tokenize
+from tokenize import untokenize, NUMBER, STRING, NAME, OP
+
+logger = logging.getLogger(__name__)
+
+if hasattr(logging, "NullHandler"):
+    logger.addHandler(logging.NullHandler())
+else:
+
+    class NullHandler(logging.Handler):
+        """
+        This handler does nothing. It's intended to be used to avoid the
+        "No handlers could be found for logger XXX" one-off warning. This is
+        important for library code, which may contain code to log events. If a user
+        of the library does not configure logging, the one-off warning might be
+        produced; to avoid this, the library developer simply needs to instantiate
+        a NullHandler and add it to the top-level logger of the library module or
+        package.
+        """
+
+        def handle(self, record):
+            pass
+
+        def emit(self, record):
+            pass
+
+        def createLock(self):
+            self.lock = None
+
+    logger.addHandler(NullHandler())
+
+if hasattr(functools, "total_ordering"):
+    total_ordering = functools.total_ordering
+else:
+
+    def total_ordering(cls):
+        """Class decorator that fills in missing ordering methods"""
+        convert = {
+            "__lt__": [
+                ("__gt__", lambda self, other: not (self < other or self == other)),
+                ("__le__", lambda self, other: self < other or self == other),
+                ("__ge__", lambda self, other: not self < other),
+            ],
+            "__le__": [
+                ("__ge__", lambda self, other: not self <= other or self == other),
+                ("__lt__", lambda self, other: self <= other and not self == other),
+                ("__gt__", lambda self, other: not self <= other),
+            ],
+            "__gt__": [
+                ("__lt__", lambda self, other: not (self > other or self == other)),
+                ("__ge__", lambda self, other: self > other or self == other),
+                ("__le__", lambda self, other: not self > other),
+            ],
+            "__ge__": [
+                ("__le__", lambda self, other: (not self >= other) or self == other),
+                ("__gt__", lambda self, other: self >= other and not self == other),
+                ("__lt__", lambda self, other: not self >= other),
+            ],
+        }
+        roots = set(dir(cls)) & set(convert)
+        if not roots:
+            raise ValueError("must define at least one ordering operation: < > <= >=")
+        root = max(roots)  # prefer __lt__ to __le__ to __gt__ to __ge__
+        for opname, opfunc in convert[root]:
+            if opname not in roots:
+                # print (opname, opfunc)
+                opfunc.__name__ = opname
+                opfunc.__doc__ = getattr(int, opname).__doc__
+                setattr(cls, opname, opfunc)
+        return cls
+
+
+if sys.version < "3":
+    from io import open
[…]

[…]
