+++ pyphot/future/pbar.py
@@ -1,0 +1,327 @@
+"""
+Simple progressbar
+==================
+
+This package implement a unique progress bar class that can be used to decorate
+an iterator, a function or even standalone.
+
+The format of the meter is flexible and can display along with the progress
+meter, the running time, an eta, and the rate of the iterations.
+
+An example is::
+    description    [----------] k/n  10% [time: 00:00:00, eta: 00:00:00, 2.7 iters/sec]
+"""
+
+import time as _time
+import sys
+import signal
+from array import array
+from fcntl import ioctl
+import termios
+
+
+__all__ = ["Pbar"]
+
+
+class Pbar(object):
+    """
+    make a progress string  in a shape of::
+
+        [----------] k/n  10% [time: 00:00:00, eta: 00:00:00, 2.7 iters/sec]
+
+
+    Attributes
+    ----------
+
+    time: bool, optional (default: True)
+        if set, add the runtime information
+
+    eta: bool, optional (default: True)
+        if set, add an estimated time to completion
+
+    rate: bool, optional (default: True)
+        if set, add the rate information
+
+    length: int, optional (default: None)
+        number of characters showing the progress meter itself
+        if None, the meter will adapt to the buffer width
+
+        TODO: make it variable with the buffer length
+
+    keep: bool, optional (default: True)
+        If not set, deletes its traces from screen after completion
+
+    file: buffer
+        the buffer to write into
+
+    mininterval: float (default: 0.5)
+        minimum time in seconds between two updates of the meter
+
+    miniters: int, optional (default: 1)
+        minimum iteration number between two updates of the meter
+
+    units: str, optional (default: 'iters')
+        unit of the iteration
+    """
+
+    def __init__(
+        self,
+        maxval=None,
+        desc=None,
+        time=True,
+        eta=True,
+        rate=True,
+        length=None,
+        file=None,
+        keep=True,
+        mininterval=0.5,
+        miniters=1,
+        units="iters",
+        **kwargs,
+    ):
+        self.time = time
+        self.eta = eta
+        self.rate = rate
+        self.desc = desc or ""
+        self.units = units
+        self.file = file or sys.stdout
+        self._last_print_len = 0
+        self.keep = keep
+        self.mininterval = mininterval
+        self.miniters = miniters
+        self._auto_width = True
+        self.length = 10
+        if length is not None:
+            self.length = length
+            self._auto_width = False
+        # backward compatibility
+        self._start_t = _time.time()
+        self._maxval = maxval
+        if "txt" in kwargs:
+            self.desc = kwargs["txt"]
+
+    def _buffer_width(self):
+        """returns the width of the buffer when available"""
+        try:
+            self.handle_resize(None, None)
+            signal.signal(signal.SIGWINCH, self.handle_resize)
+            self._auto_width = True
+        except Exception:
+            self.term_width = 79
+            self._auto_width = False
+
+        return self.term_width
+
+    def handle_resize(self, signum, frame):
+        ioinfo = ioctl(self.file, termios.TIOCGWINSZ, "\0" * 8)  # type: ignore
+        h, w = array("h", ioinfo)[:2]
+        self.term_width = w
+
+    @staticmethod
+    def format_interval(t):
+        """make a human readable time interval decomposed into days, hours,
+        minutes and seconds
+
+        Parameters
+        ----------
+        t: int
+            interval in seconds
+
+        Returns
+        -------
+        txt: str
+            string representing the interval
+            (format:  <days>d <hrs>:<min>:<sec>)
+        """
+        mins, s = divmod(int(t), 60)
+        h, m = divmod(mins, 60)
[…]
