### 2025-10-31 b16d081
add unit tests for future.
