### 2025-11-03 6b2b991
add lick testing
