### 2025-11-03 7d3aca1
add licks features
