+++ pyphot/future/licks.py
@@ -1,0 +1,631 @@
+"""Lick indices calculations
+
+
+This package provides function to compute spectral indices
+
+
+A collection of many common indices is available in `licks.dat`
+
+The Lick system of spectral line indices is one of the most commonly used
+methods of determining ages and metallicities of unresolved (integrated light)
+stellar populations.
+
+The calibration of the Lick/ IDS system is complicated because the original
+Lick spectra were not flux calibrated, so there are usually systematic effects
+due to differences in continuum shape.  Proper calibration involves observing
+many of the original Lick/IDS standard stars and deriving offsets to the
+standard system.
+
+references
+~~~~~~~~~~
+
+    Worthey G., Faber S. M., Gonzalez J. J., Burstein D., 1994, ApJS, 94, 687
+    Worthey G., Ottaviani D. L., 1997, ApJS, 111, 377
+    Puzia et al. 2002
+    Zhang, Li & Han 2005, http://arxiv.org/abs/astro-ph/0508634v1
+
+
+notes
+~~~~~
+
+    In Vazdekis et al. (2010), we propose a new Line Index System, hereafter
+    LIS, with three new spectral resolutions at which to measure the Lick
+    indices. Note that this new system should not be restricted to the Lick set
+    of indices in a flux calibrated system. In fact, LIS can be used for any
+    index in the literature (e.g., for the Rose (1984) indices), including
+    newly defined indices (e.g., Cervantes & Vazdekis 2009).
+
+
+    The LIS system is defined for 3 different spectral resolutions which are
+    best suited for the following astrophysical cases:
+
+    LIS-5.0AA: globular clusters
+    LIS-8.4AA: low and intermediate-mass galaxies
+    LIS-14.0AA: massive galaxies
+    Conversions to transform the data from the Lick/IDS system to LIS can be
+    found
+
+
+    discussion of indices and information
+    Johansson, Thomas & Maraston 2010
+    http://wwwmpa.mpa-garching.mpg.de/~jonasj/milesff/milesff.pdf
+
+"""
+
+from dataclasses import dataclass
+from typing import Union, Literal, Optional, Tuple, Any, Dict, List
+import numpy as np
+import numpy.typing as npt
+
+from scipy.integrate import trapezoid
+from scipy.interpolate import InterpolatedUnivariateSpline
+
+from .unit_adapters import QuantityType, enforce_default_units
+from . import config
+
+
+def _reduce_resolution(
+    wi_AA: npt.NDArray[np.floating],
+    fi_flam: npt.NDArray[np.floating],
+    fwhm0_AA: float = 0.55,
+    sigma_floor_AA: float = 0.2,
+) -> npt.NDArray[np.floating]:
+    """Adapt the resolution of the spectra to match the lick definitions
+
+    Lick definitions have different resolution elements as function
+    of wavelength. These definition are hard-coded in this function
+
+    Parameters
+    ----------
+    wi_AA: ndarray (n, )
+        wavelength definition in Angstroms
+    fi_flam: ndarray (nspec, n) or (n, )
+        spectra to convert in flam
+    fwhm0_AA: float
+        initial broadening in the spectra `fi` in Angstroms
+    sigma_floor_AA: float
+        minimal dispersion to consider in Angstroms
+
+    Returns
+    -------
+    flux_red: ndarray (nspec, n) or (n, )
+        reduced spectra in flam
+    """
+
+    # all in AA
+    w_lick_res = (4000.0, 4400.0, 4900.0, 5400.0, 6000.0)
+    lick_res = (11.5, 9.2, 8.4, 8.4, 9.8)  # FWHM in AA
+
+    w = np.asarray(wi_AA)
[…]
