### 2025-11-03 740cd9d
fix expand syntax on older python
