### 2025-11-03 d42cab0
fix union syntax error on python <3.12
