+++ unittests/test_future_svo.py
@@ -1,0 +1,18 @@
+"""Test the future SVO filter interface"""
+
+from pyphot.future.svo import get_pyphot_filter
+from pyphot.future.phot import Filter
+from pyphot.future import config
+
+
+def test_svo_filter_interface(unit_backend=None):
+    """Test the future SVO filter interface"""
+    lst = "2MASS/2MASS.J 2MASS/2MASS.H 2MASS/2MASS.Ks HST/ACS_WFC.F475W HST/ACS_WFC.F814W".split()
+    if unit_backend is not None:
+        config.set_units_backend(unit_backend)
+
+    for name in lst:
+        pb = get_pyphot_filter(name)
+        pb.info()
+        assert isinstance(pb, Filter)
+        assert pb.name == name.replace("/", "_")
