+++ pyphot/future/unit_adapters/units_adapter.py
@@ -10,8 +10,6 @@ from typing import Any, Optional, Callable, Sequence, cast, Union
 from functools import wraps
 import warnings
 import inspect
-import inspect
-from functools import wraps
 from textwrap import dedent
 
 
