### 2025-11-04 6d80c11
add py 3.14 in test versions
