### 2025-11-04 ab92914
updated classifier for python versions
