+++ pyproject.toml
@@ -14,7 +14,11 @@ authors = [{name="Morgan Fouesneau"}]
 classifiers=[
       'Intended Audience :: Science/Research',
       'Operating System :: OS Independent',
-      'Programming Language :: Python :: 3',
+      'Programming Language :: Python :: 3.9',
+      'Programming Language :: Python :: 3.10',
+      'Programming Language :: Python :: 3.11',
+      'Programming Language :: Python :: 3.12',
+      'Programming Language :: Python :: 3.13',
       'License :: OSI Approved :: MIT License',
       'Topic :: Scientific/Engineering :: Astronomy'
       ]
