### 2025-11-04 3c4e68b
removed py3.14 tables blocking [hdf5 library]
