### 2025-11-05 e90133a
add / fix ascii readers
