+++ unittests/test_future_library.py
@@ -7,9 +7,28 @@ from typing import cast
 
 from pyphot.future.libraries import HDF_Library, Ascii_Library, get_library
 from pyphot import config
-from pyphot.future.phot import Filter
+from pyphot.future.phot import _drop_units, Filter
 
 lib = get_library()
+
+reference_info = {
+    "GROUND_JOHNSON_V": {
+        "dtype": "energy",
+        "wavelength_unit": "AA",
+        "cl": 5504.666021,
+        "lpivot": 5469.853351,
+        "leff": 5438.689749,
+        "lphot": 5460.082104,
+        "lmin": 4766.670000,
+        "lmax": 6899.950000,
+        "norm": 870.650149,
+        "width": 870.651020,
+        "fwhm": 827.370000,
+        "Vega_zero_mag": 21.099102,
+        "AB_zero_mag": 21.097827,
+        "ST_zero_mag": 21.100000,
+    }
+}
 
 
 def test_hdf_library():
@@ -52,3 +71,24 @@ def test_integrity_in_internal_library(name):
         fk.ST_zero_Jy.value,
     )
     print(fmt.format(*rec))  # checks value types
+
+
+def test_ascii_library():
+    """Test ascii library using internal files"""
+    where = pathlib.Path(str(config.__default_passband_lib__)).parent
+    where = str(where / "ascii_sources")
+    lib = Ascii_Library(where)
+    assert isinstance(lib, Ascii_Library)
+    assert lib.source is not None
+    assert len(lib) > 0
+    which = lib.find("johnson_v")
+    assert "ground_johnson_v" in which[0]
+    pb = lib.load_filters(which)[0]
+
+    refs = reference_info["GROUND_JOHNSON_V"]
+    for key, value in refs.items():
+        other = _drop_units(getattr(pb, key))
+        if isinstance(value, str):
+            assert value == other
+        else:
+            assert abs(value - other) < 1e-6
