+++ pyphot/future/io/ascii.py
@@ -1,0 +1,588 @@
+"""Export dataframe to ASCII format while preserving attrs"""
+
+import pandas as pd
+from typing import Callable, Hashable, Sequence, List, Tuple, cast, Union
+from pandas.io.common import get_handle
+from pandas._typing import (
+    CompressionOptions,
+    BaseBuffer,
+    FilePath,
+    IndexLabel,
+    StorageOptions,
+    OpenFileErrors,
+)
+from io import IOBase
+from os import PathLike
+
+from typing import Any, Dict, Hashable
+from dataclasses import dataclass
+
+
+@dataclass
+class HeaderInfo:
+    """Extracted information from FITS header
+
+    Attributes
+    ----------
+        header: dict
+            header dictionary
+
+        alias: dict
+            aliases
+
+        units: dict
+            units
+
+        comments: dict
+            comments/description of keywords
+    """
+
+    header: Dict[Hashable, Any]
+    alias: Dict[Hashable, str]
+    units: Dict[Hashable, str]
+    comments: Dict[Hashable, str]
+
+
+def ascii_read_header(
+    fname: str | FilePath | IOBase,
+    *,
+    commentchar: str = "#",
+    delimiter: str = ",",
+    commented_header: bool = True,
+    **kwargs,
+) -> Tuple[int, HeaderInfo, List[str]]:
+    """
+    Read ASCII/CSV header
+
+    Parameters
+    ----------
+    fname: str, FilePath, BaseBuffer
+        File, filename, or generator to read.
+        Note that generators should return byte strings for Python >=3.
+
+    comments: str, optional
+        The character used to indicate the start of a comment;
+        default: '#'. ("" is equivalent to None)
+
+    delimiter: str, optional
+        The string used to separate values.  By default, this is any
+        whitespace.
+
+    commented_header: bool, optional
+        if set, the last line of the header is expected to be the column titles (with comment character)
+        otherwise, the first line of the data will be the column titles
+
+    Returns
+    -------
+    nlines: int
+        number of lines from the header
+
+    info: HeaderInfo
+        header information (header, alias, units, comments)
+
+    names: List[str]
+        sequence or str, first data line after header, expected to be the
+        column names.
+    """
+
+    # define some internal functions
+    def parseStrNone(v: str) -> str | None:
+        """robust parse"""
+        _v = v.strip().split()
+        if len(_v) == 0:
+            return None
+        else:
+            _v = " ".join(_v)
+            if (_v.lower()) == "none" or (_v.lower() == "null"):
+                return None
+            return _v
+
+    def parseColInfo(line: str) -> tuple[str, str | None, str | None]:
+        """parse column info"""
+        line = line.replace(commentchar, "").strip()
+        tokens = line.split("\t")
+        colname = tokens[0].strip()
+        if len(tokens) > 1:
+            colunit = parseStrNone(tokens[1].strip())
+        else:
+            colunit = None
+        if len(tokens) > 2:
+            colcomm = tokens[2].strip()
+        else:
+            colcomm = None
+        return colname, colunit, colcomm
+
+    if hasattr(fname, "read"):
+        stream: IOBase = cast(IOBase, fname)
+    else:
+        stream: IOBase = open(fname, "r")  # pyright: ignore
+
+    if commentchar is None:
+        commentchar = "#"
+
+    # initialize storage
+    alias = {}
+    units = {}
+    desc = {}
+    header = {}
+    comment = []
+    history = []
+
+    done = False
+    line = ""
+    oldline = ""  # contains the last processed line
+    nlines = 0  # contains the total number of lines processes in the header
+    names = []  # contains the names of the columns
+
+    while not done:
+        line = str(stream.readline().rstrip())  # getting rid of '\n'
+        nlines += 1
+
[…]
