+++ pyphot/future/io/__init__.py
@@ -1,0 +1,1 @@
+"""Adapted from SimpleTable (v2.0; https://github.com/mfouesneau/simpletable) with minimal depedencies"""

[…]
