+++ pyphot/sun.py
@@ -3,16 +3,15 @@
 from __future__ import print_function
 
 import warnings
+from typing import Optional, Literal, Tuple, cast
+import numpy.typing as npt
+import pandas as pd
+
+from . import config
+from .unit_adapters import QuantityType
+from . import io
 
 from .config import libsdir
-from .ezunits import unit
-from .simpletable import SimpleTable
-
-try:
-    from astropy.io import fits as pyfits
-except ImportError:
-    import pyfits  # noqa
-
 
 __all__ = ["Sun"]
 
@@ -20,88 +19,124 @@ _default_sun = {
     "observed": "{0}/sun_reference_stis_001.fits".format(libsdir),
     "theoretical": "{0}/sun_kurucz93.fits".format(libsdir),
 }
-_default_distance = "1 * au"
 
 
-class Sun(object):
+def get_library_default_distance() -> QuantityType:
+    """Return the default distance associated with the library spectra of the Sun."""
+    return 1.0 * config.units.U("au")
+
+
+class Sun:
     """
-    Class that handles the Sun's spectrum and references.
+     Class that handles the Sun's spectrum and references.
 
-    Observed solar spectrum comes from:
-    ftp://ftp.stsci.edu/cdbs/current_calspec/sun_reference_stis_001.fits
+     Observed solar spectrum comes from:
+     ftp://ftp.stsci.edu/cdbs/current_calspec/sun_reference_stis_001.fits
 
-    and theoretical spectrum comes from:
-    ftp://ftp.stsci.edu/cdbs/grid/k93models/standards/sun_kurucz93.fits
+     and theoretical spectrum comes from:
+     ftp://ftp.stsci.edu/cdbs/grid/k93models/standards/sun_kurucz93.fits
 
-    The theoretical spectrum is scaled to match the observed spectrum from 1.5 -
-    2.5 microns, and then it is used where the observed spectrum ends.
-    The theoretical model of the Sun from Kurucz'93 atlas using the following
-    parameters when the Sun is at 1 au.
+     The theoretical spectrum is scaled to match the observed spectrum from 1.5 -
+     2.5 microns, and then it is used where the observed spectrum ends.
+     The theoretical model of the Sun from Kurucz'93 atlas using the following
+     parameters when the Sun is at 1 au.
 
-     log_Z         T_eff        log_g           V_{Johnson}
-     +0.0           5777        +4.44              -26.75
+    +-----------+------------+----------+-------------+
+    | log(Z)    | Teff       | log_g    | V_{Johnson} |
+    +===========+============+==========+=============+
+    | +0.0      | +5777      | +4.44    | -26.75      |
+    +-----------+------------+----------+-------------+
 
-    Attributes
-    ----------
-    source: str
-        filename of the sun library
-    data: SimpleTable
-        data table
-    units: tuple
-        detected units from file header
-    distance: float
-        distance to the observed Sun (default, 1 au)
-    flavor: str, (default theoretical)
-        either 'observed' using the stis reference,
-        or  'theoretical' for the Kurucz model.
+     Attributes
+     ----------
+     source: str
+         filename of the sun library
+     data: SimpleTable
+         data table
+     units: tuple
+         detected units from file header
+     wavelength: array
+         wavelength (with units when found)
+     flux: array
+         flux(wavelength) values (with units when provided)
+     distance: float
+         distance to the observed Sun (default, 1 au)
+     flavor: str, (default theoretical)
+         either 'observed' using the stis reference,
+         or  'theoretical' for the Kurucz model.
     """
 
-    def __init__(self, source=None, distance=1 * unit["au"], flavor="theoretical"):
+    _data: Optional[pd.DataFrame]
+    units: Optional[Tuple[str, str]]
+    distance: QuantityType
+    distance_conversion: float
[…]
