+++ pyphot/io/fits.py
@@ -1,0 +1,279 @@
+from typing import Any, Tuple
+from astropy.io import fits
+from astropy.io.fits import Header
+from astropy.io.fits import TableHDU, BinTableHDU
+import numpy.typing as npt
+import pandas as pd
+from .header import HeaderInfo
+
+
+try:
+    from astropy.io import fits
+
+    def fix_endian_issue(arr: npt.NDArray | Any) -> npt.NDArray:
+        """Fix endian issue in array which happens often when reading FITS files"""
+        return arr.astype(arr.dtype.newbyteorder())
+
+    def fits_read_header(hdr: Header) -> HeaderInfo:
+        """
+        Convert pyfits header into dictionary with relevant values
+
+        Parameters
+        ----------
+
+        hdr: pyftis.Header
+            fits unit
+
+        Returns
+        -------
+        headerinfo: HeaderInfo
+            extracted information from header
+        """
+        header = {}
+        alias = {}
+        units = {}
+        comments = {}
+
+        # generic cards
+        genTerms = [
+            "XTENSION",
+            "BITPIX",
+            "NAXIS",
+            "NAXIS1",
+            "NAXIS2",
+            "PCOUNT",
+            "GCOUNT",
+            "TFIELDS",
+            "EXTNAME",
+        ]
+        fieldTerms = ["TTYPE", "TFORM", "TUNIT", "ALIAS"]
+
+        for card in hdr.cards["TTYPE*"]:
+            name = card.value
+            comments[name] = card.comment
+            u = hdr.get(card.keyword.replace("TYPE", "UNIT"), None)
+            if u is not None:
+                units[name] = u
+
+        # for k, val, _ in hdr.ascard['ALIAS*']:
+        for card in hdr.cards["ALIAS*"]:
+            k = card.keyword
+            val = card.value
+            al, orig = val.split("=")
+            alias[al] = orig
+
+        # other specific keywords: COMMENT, HISTORY
+        header_comments = []
+        header_history = []
+        for k, v in hdr.items():
+            if (k not in genTerms) and (k[:5] not in fieldTerms):
+                if k == "COMMENT":
+                    header_comments.append(v)
+                elif k == "HISTORY":
+                    header_history.append(v)
+                else:
+                    header[k] = v
+
+        # COMMENT, HISTORY polish
+        if len(header_comments) > 0:
+            header["COMMENT"] = "\n".join(header_comments)
+        if len(header_history) > 0:
+            header["HISTORY"] = "\n".join(header_history)
+
+        if "EXTNAME" in hdr:
+            header["NAME"] = hdr["EXTNAME"]
+
+        return HeaderInfo(header, alias, units, comments)
+
+    def from_fits(
+        filename: str, extension_number: int = 1
+    ) -> Tuple[npt.NDArray, HeaderInfo]:
+        """Load a DataFrame from a FITS file.
+
+        Parameters
+        ----------
+        filename : str
+            The path to the FITS file.
+        extension_number : int, optional
+            The extension number to load, by default 1.
+
+        Returns
+        -------
+        Tuple[npt.NDArray, HeaderInfo]
+            The loaded DataFrame and its header information.
+        """
+        with fits.open(filename) as hdu:
+            extension: TableHDU | BinTableHDU = hdu[extension_number]  # pyright: ignore[reportAssignmentType]
+            hdr_info = fits_read_header(extension.header)
+            data = fix_endian_issue(extension.data)
+        return data, hdr_info
+
+except ImportError:
+    from_fits = None  # pyright: ignore[reportAssignmentType]
+
+
+def fits_generate_header(df: pd.DataFrame) -> fits.Header:
[…]

[…]
