### 2025-11-05 bf2dbb0
fix py3.9 union
