### 2025-11-06 1c32cfa
fix distance unit simplification not automatic
