### 2025-11-06 be30c1e
fix literalString issue with py <3.11
