### 2025-11-06 5fc4972
fix test when missing unit packages
