### 2025-11-06 c1d99fb
Fix Union for py < 3.10
