### 2025-11-06 aa12704
fix py<3.10 union typing
