### 2025-11-06 489132b
fix py <3.10 typing union
