### 2025-11-06 042e46d
fix py<3.10 union typing
