### 2025-11-06 966d6e0
minor fix edge case
