### 2025-11-06 474389e
split and improved test coverage
