+++ unittests/test_sun.py
@@ -1,0 +1,406 @@
+"""testing the vega/sun interfaces"""
+
+import pytest
+import os
+import tempfile
+import pandas as pd
+from unittest.mock import patch, Mock, mock_open
+
+from typing import cast
+import numpy as np
+
+from pyphot.libraries import get_library
+from pyphot import sun, config
+from pyphot.phot import Filter
+from pyphot.unit_adapters import backends
+
+
+@pytest.mark.parametrize("flavor", list(sun._default_sun.keys()))
+def test_instanciate_sun(flavor):
+    """Checks that all registered flavors can be instantiated"""
+    s = sun.Sun(flavor=flavor)
+    assert s.data is not None
+    assert s._data is not None  # should not be after the previous call
+    assert s.data is not None
+
+
+@pytest.mark.parametrize("AdapterName", list(backends.keys()))
+def test_sun_magnitudes(AdapterName):
+    if AdapterName is not None:
+        try:
+            config.set_units_backend(AdapterName)
+        except ImportError:
+            pytest.skip(f"Skipping test for {AdapterName} adapter")
+    sun_obs = sun.Sun(flavor="observed")
+    sun_th = sun.Sun()  # default is theoric spectrum
+    sun_th_10pc = sun.Sun(distance=10 * config.units.U("pc"))
+
+    lib = get_library()
+
+    f = cast(Filter, lib["GROUND_JOHNSON_V"])
+    expectations = -26.76, -26.76, +4.81
+
+    for name, sun_flavor, expect_mag in zip(
+        ("observed", "theoretical", "th. 10pc"),
+        (sun_obs, sun_th, sun_th_10pc),
+        expectations,
+    ):
+        flux = f.get_flux(sun_flavor.wavelength, sun_flavor.flux)
+        vegamag = f.Vega_zero_mag
+        sun_vega_mag = -2.5 * np.log10(flux.value) - vegamag
+        assert abs(sun_vega_mag - expect_mag) < 1e-2
+
+
+def test_get_library_default_distance():
+    """Test the get_library_default_distance function"""
+    distance = sun.get_library_default_distance()
+
+    # Should return 1 AU
+    assert distance.value == 1.0
+    # Should have appropriate units (AU)
+    assert str(distance.unit).lower() in ["au", "astronomical_unit"]
+
+
+def test_sun_constructor_defaults():
+    """Test Sun constructor with default parameters"""
+    sun_obj = sun.Sun()
+
+    assert sun_obj.flavor == "theoretical"
+    assert sun_obj.source == sun._default_sun["theoretical"]
+    assert sun_obj._data is None
+    assert sun_obj.units is None
+    assert sun_obj.distance_conversion == 1.0
+
+
+def test_sun_constructor_with_flavor():
+    """Test Sun constructor with different flavors"""
+    # Test theoretical flavor
+    sun_theo = sun.Sun(flavor="theoretical")
+    assert sun_theo.flavor == "theoretical"
+    assert sun_theo.source == sun._default_sun["theoretical"]
+
+    # Test observed flavor
+    sun_obs = sun.Sun(flavor="observed")
+    assert sun_obs.flavor == "observed"
+    assert sun_obs.source == sun._default_sun["observed"]
+
+
+def test_sun_constructor_with_custom_source():
+    """Test Sun constructor with custom source file"""
+    custom_source = "/path/to/custom_sun.fits"
+    sun_obj = sun.Sun(source=custom_source, flavor="observed")
+
[…]
