+++ unittests/test_vega.py
@@ -1,0 +1,385 @@
+"""testing the vega/sun interfaces"""
+
+import pytest
+import os
+import tempfile
+import pandas as pd
+from unittest.mock import patch, Mock, mock_open
+
+from typing import cast
+import numpy as np
+
+from pyphot.libraries import get_library
+from pyphot import vega, config
+from pyphot.phot import Filter
+from pyphot.unit_adapters import backends
+
+
+@pytest.mark.parametrize("flavor", list(vega._default_vega.keys()))
+def test_instanciate_vega(flavor):
+    """Checks that all registered flavors can be instantiated"""
+    v = vega.Vega(flavor=flavor)
+    assert v.data is not None
+    assert v._data is not None  # should not be after the previous call
+    assert v.data is not None  # checks the cache
+
+
+def test_vega_constructor_defaults():
+    """Test Vega constructor with default parameters"""
+    vega_obj = vega.Vega()
+
+    assert vega_obj.source == vega._default_vega["legacy"]
+    assert vega_obj._data is None
+    assert vega_obj.units is None
+
+
+def test_vega_constructor_with_flavor():
+    """Test Vega constructor with different flavors"""
+    # Test with each available flavor
+    for flavor in vega._default_vega.keys():
+        vega_obj = vega.Vega(flavor=flavor)
+        assert vega_obj.source == vega._default_vega[flavor]
+
+
+def test_vega_constructor_with_custom_source():
+    """Test Vega constructor with custom source file"""
+    custom_source = "/path/to/custom_vega.fits"
+    vega_obj = vega.Vega(source=custom_source)
+
+    assert vega_obj.source == custom_source
+
+
+def test_vega_constructor_invalid_flavor():
+    """Test Vega constructor with invalid flavor"""
+    with pytest.raises(ValueError, match="Unknown Vega flavor"):
+        vega.Vega(flavor="invalid_flavor")
+
+
+def test_vega_constructor_no_source_no_flavor():
+    """Test Vega constructor with neither source nor flavor"""
+    with pytest.raises(
+        RuntimeError, match="Either `source` or `flavor` must be provided"
+    ):
+        vega.Vega(source=None, flavor=None)  # type: ignore / on purpose
+
+
+def test_set_source_flavor_with_source():
+    """Test _set_source_flavor method with custom source"""
+    vega_obj = vega.Vega()
+    custom_source = "/custom/path.fits"
+
+    vega_obj._set_source_flavor(source=custom_source, flavor=None)
+
+    assert vega_obj.source == custom_source
+    assert vega_obj._data is None
+    assert vega_obj.units is None
+
+
+def test_set_source_flavor_with_flavor():
+    """Test _set_source_flavor method with flavor only"""
+    vega_obj = vega.Vega(flavor="legacy")  # Start with legacy
+
+    # Change to mod_003
+    vega_obj._set_source_flavor(source=None, flavor="mod_003")
+
+    assert vega_obj.source == vega._default_vega["mod_003"]
+
+
+def test_set_source_flavor_invalid_flavor():
+    """Test _set_source_flavor method with invalid flavor"""
+    vega_obj = vega.Vega()
+
+    with pytest.raises(ValueError, match="Unknown Vega flavor"):
+        vega_obj._set_source_flavor(source=None, flavor="invalid")
+
+
+def test_set_source_flavor_no_parameters():
[…]
