+++ unittests/test_pbar.py
@@ -1,0 +1,461 @@
+"""Test pbar module functionality"""
+
+import pytest
+import sys
+import io
+import time
+from unittest.mock import Mock, patch, MagicMock, call
+from array import array
+
+from pyphot.pbar import Pbar
+
+
+def test_pbar_init_defaults():
+    """Test Pbar initialization with default parameters"""
+    pbar = Pbar()
+
+    assert pbar.time is True
+    assert pbar.eta is True
+    assert pbar.rate is True
+    assert pbar.desc == ""
+    assert pbar.units == "iters"
+    assert pbar.file == sys.stdout
+    assert pbar._last_print_len == 0
+    assert pbar.keep is True
+    assert pbar.mininterval == 0.5
+    assert pbar.miniters == 1
+    assert pbar._auto_width is True
+    assert pbar.length == 10
+    assert pbar._maxval is None
+
+
+def test_pbar_init_custom_parameters():
+    """Test Pbar initialization with custom parameters"""
+    custom_file = io.StringIO()
+    pbar = Pbar(
+        maxval=100,
+        desc="Testing",
+        time=False,
+        eta=False,
+        rate=False,
+        length=20,
+        file=custom_file,
+        keep=False,
+        mininterval=1.0,
+        miniters=5,
+        units="items",
+    )
+
+    assert pbar._maxval == 100
+    assert pbar.desc == "Testing"
+    assert pbar.time is False
+    assert pbar.eta is False
+    assert pbar.rate is False
+    assert pbar.length == 20
+    assert pbar.file == custom_file
+    assert pbar.keep is False
+    assert pbar.mininterval == 1.0
+    assert pbar.miniters == 5
+    assert pbar.units == "items"
+    assert pbar._auto_width is False
+
+
+def test_pbar_init_backward_compatibility():
+    """Test Pbar initialization with backward compatibility txt parameter"""
+    pbar = Pbar(txt="Legacy description")
+    assert pbar.desc == "Legacy description"
+
+
+def test_format_interval():
+    """Test the format_interval static method"""
+    # Test seconds only
+    assert Pbar.format_interval(45) == "00:45"
+
+    # Test minutes and seconds
+    assert Pbar.format_interval(125) == "02:05"
+
+    # Test hours, minutes, and seconds
+    assert Pbar.format_interval(3665) == "01:01:05"
+
+    # Test days, hours, minutes, and seconds
+    assert Pbar.format_interval(90061) == "1d 01:01:01"
+
+    # Test zero
+    assert Pbar.format_interval(0) == "00:00"
+
+    # Test large values
+    assert Pbar.format_interval(172861) == "2d 01:01"
+
+
+def test_build_str_meter_no_total():
+    """Test build_str_meter with no total (indeterminate progress)"""
+    pbar = Pbar(time=True, eta=True, rate=True)
+
+    # Use 0 as total to trigger the "not total" behavior
+    result = pbar.build_str_meter(50, 0, 10.5)
+
+    assert "50" in result
+    assert "00:10" in result  # elapsed time
+    assert "4.76 iters/sec" in result  # rate
+    assert "|" not in result  # no progress bar for indeterminate
+
+
+def test_build_str_meter_with_total():
+    """Test build_str_meter with known total"""
+    pbar = Pbar(time=True, eta=True, rate=True, length=10)
+
+    result = pbar.build_str_meter(25, 100, 10.0)
+
+    assert "|" in result
+    assert "25/100" in result
+    assert "25%" in result
+    assert "##--" in result  # progress bar (25% of 10 chars = 2.5 ≈ 2 chars)
+    assert "00:10" in result  # elapsed time
[…]
