+++ unittests/test_helpers.py
@@ -1,0 +1,182 @@
+"""Test helpers module functionality"""
+
+import pytest
+import warnings
+from unittest.mock import Mock, patch
+from math import pi
+
+from pyphot.helpers import progress_enumerate, deprecated, distc
+
+
+class TestProgressEnumerate:
+    """Test the progress_enumerate function"""
+
+    def test_progress_enumerate_without_progress(self):
+        """Test progress_enumerate with show_progress=False"""
+        items = [1, 2, 3, 4, 5]
+        result = list(progress_enumerate(items, show_progress=False))
+        expected = [(0, 1), (1, 2), (2, 3), (3, 4), (4, 5)]
+        assert result == expected
+
+    def test_progress_enumerate_default_no_progress(self):
+        """Test progress_enumerate with default parameters (no progress)"""
+        items = ["a", "b", "c"]
+        result = list(progress_enumerate(items))
+        expected = [(0, "a"), (1, "b"), (2, "c")]
+        assert result == expected
+
+    def test_progress_enumerate_with_start_parameter(self):
+        """Test progress_enumerate with start parameter"""
+        items = [10, 20, 30]
+        result = list(progress_enumerate(items, 5, show_progress=False))
+        expected = [(5, 10), (6, 20), (7, 30)]
+        assert result == expected
+
+    @patch("pyphot.helpers.Pbar")
+    def test_progress_enumerate_with_progress(self, mock_pbar_class):
+        """Test progress_enumerate with show_progress=True"""
+        # Setup mock
+        mock_pbar_instance = Mock()
+        mock_pbar_class.return_value = mock_pbar_instance
+        mock_pbar_instance.iterover.return_value = [1, 2, 3]
+
+        items = [1, 2, 3]
+        result = list(progress_enumerate(items, show_progress=True))
+        expected = [(0, 1), (1, 2), (2, 3)]
+
+        assert result == expected
+        mock_pbar_class.assert_called_once()
+        mock_pbar_instance.iterover.assert_called_once_with(items)
+
+    @patch("pyphot.helpers.Pbar")
+    def test_progress_enumerate_with_progress_and_kwargs(self, mock_pbar_class):
+        """Test progress_enumerate with show_progress=True and additional kwargs"""
+        # Setup mock
+        mock_pbar_instance = Mock()
+        mock_pbar_class.return_value = mock_pbar_instance
+        mock_pbar_instance.iterover.return_value = ["x", "y"]
+
+        items = ["x", "y"]
+        result = list(progress_enumerate(items, show_progress=True, desc="Testing"))
+        expected = [(0, "x"), (1, "y")]
+
+        assert result == expected
+        mock_pbar_class.assert_called_once_with(desc="Testing")
+        mock_pbar_instance.iterover.assert_called_once_with(items)
+
+    def test_progress_enumerate_empty_sequence(self):
+        """Test progress_enumerate with empty sequence"""
+        items = []
+        result = list(progress_enumerate(items, show_progress=False))
+        assert result == []
+
+    def test_progress_enumerate_single_item(self):
+        """Test progress_enumerate with single item"""
+        items = [42]
+        result = list(progress_enumerate(items, show_progress=False))
+        expected = [(0, 42)]
+        assert result == expected
+
+
+class TestDeprecatedDecorator:
+    """Test the deprecated decorator"""
+
+    def test_deprecated_decorator_warns(self):
+        """Test that deprecated decorator issues a DeprecationWarning"""
+        test_message = "This function is deprecated"
+
+        @deprecated(test_message)
+        def test_function():
+            return "test_result"
+
+        with warnings.catch_warnings(record=True) as w:
+            warnings.simplefilter("always")
+            result = test_function()
+
[…]

[…]
