+++ .github/workflows/unittests.yml
@@ -23,6 +23,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install ".[testing]"
-      - name: Running quick tests
+      - name: Running tests
         run: |
           python3 -m pytest
