### 2025-11-06 e2a8c56
add coverage output to logs
