### 2025-11-06 3909752
clean unused variable
