### 2025-11-06 cb8eeaa
makes and tests svo independent from astropy
