### 2025-11-06 df84afc
add internal votable reader
