+++ pyphot/io/votable.py
@@ -1,0 +1,342 @@
+"""
+VOTable parser for astronomical tabular data.
+
+VOTable is the standard XML format for astronomical tabular data.
+This module implements a custom `VOTableParser` that uses XML parsing and not other dependencies.
+`from_votable` provides the standard interface of io operations (pandas.DataFrame, HeaderInfo)
+
+"""
+
+from io import IOBase
+from os import PathLike
+from typing import Dict, List, Any, Union, Tuple
+import numpy as np
+import numpy.typing as npt
+import pandas as pd
+import xml.etree.ElementTree as ET
+import requests
+
+from .header import HeaderInfo
+
+
+class VOTableParser:
+    """
+    A custom VOTable parser using XML parsing.
+
+    VOTable is the standard XML format for astronomical tabular data.
+    This example shows how to parse the structure and extract data.
+    """
+
+    def __init__(
+        self,
+        source: Union[str, bytes, PathLike, IOBase],
+        is_url: bool = False,
+    ):
+        """
+        Initialize VOTable parser
+
+        Parameters
+        ----------
+        source : str or bytes
+            Either a file path, URL, or XML string/bytes
+        is_url : bool
+            If True, treat source as URL to fetch
+        """
+        if is_url and isinstance(source, (str, bytes)):
+            response = requests.get(source)
+            response.raise_for_status()
+            xml_content = response.content
+        elif isinstance(source, str) and source.startswith("<?xml"):
+            xml_content = source
+        elif isinstance(source, bytes) and source.startswith(b"<?xml"):
+            xml_content = source.decode("utf-8")
+        elif isinstance(source, IOBase):
+            xml_content = source.read()
+        else:
+            # Assume it's a file path
+            with open(source, "r", encoding="utf-8") as f:
+                xml_content = f.read()
+
+        self.root = ET.fromstring(xml_content)
+        self.namespace = self._get_namespace()
+        self.tables = []
+        self._parse_tables()
+
+    def _get_namespace(self) -> Dict[str, str]:
+        """Extract namespace from VOTable root element"""
+        # VOTable typically uses namespace like http://www.ivoa.net/xml/VOTable/v1.3
+        namespace = {}
+        if self.root.tag.startswith("{"):
+            uri = self.root.tag.split("}")[0][1:]
+            namespace["vot"] = uri
+        return namespace
+
+    def _find_element(self, parent, tag: str):
+        """Find element accounting for namespace"""
+        if self.namespace:
+            return parent.find(f"{{{self.namespace['vot']}}}{tag}")
+        return parent.find(tag)
+
+    def _findall_elements(self, parent, tag: str):
+        """Find all elements accounting for namespace"""
+        if self.namespace:
+            return parent.findall(f".//{{{self.namespace['vot']}}}{tag}")
+        return parent.findall(f".//{tag}")
+
+    def _parse_tables(self):
+        """Parse all TABLE elements in the VOTable"""
+        table_elements = self._findall_elements(self.root, "TABLE")
+
+        for table_elem in table_elements:
+            table_data = self._parse_single_table(table_elem)
+            self.tables.append(table_data)
+
+    def _parse_single_table(self, table_elem) -> Dict[str, Any]:
+        """Parse a single TABLE element"""
+        table_info = {
+            "name": table_elem.get("name", "unnamed"),
+            "description": "",
+            "fields": [],
+            "params": [],
+            "data": [],
+        }
+
+        # Get table description
+        desc_elem = self._find_element(table_elem, "DESCRIPTION")
+        if desc_elem is not None:
+            table_info["description"] = desc_elem.text or ""
+
+        # Parse FIELD elements (column definitions)
+        field_elements = self._findall_elements(table_elem, "FIELD")
[…]
