+++ pyphot/README.md
@@ -2,6 +2,7 @@
 
 ## New features
 
+- major refactoring of the codebase: remove code duplication due to the split astropy vs. pint
 - add hint typing in the code (not tested with mypy; pyright ok)
 - adapters to interface transparently with varied unit frameworks (e.g., astropy.units, pint)
   - `future.config.set_units_backend` to switch between different unit frameworks
@@ -11,8 +12,8 @@
   - still available locally (since v.1.7.0)
 - added specific unittests - coverage > 70%
 - internal simpletable removed in favor of pandas
-  - python.io for ascii, fits, hdf5adapted from https://github.com/mfouesneau/simpletable
-
+  - pyphot.io for ascii, fits, hdf5, votable adapted from https://github.com/mfouesneau/simpletable
+- svo module made independent from the astropy.io.votable to math internal unit backend
 
 
 ```{warning} configuration update only affects newly created objects
@@ -27,9 +28,6 @@ If you change the vega flavor, you may have a mix of Vega definitions, which wil
 
 ## Incoming
 
+- imports arrangements
 - io.hdf does not allow export yet.
-
 - update documentation with new features/refactoring
-
-- remove dependency on astropy for the svo module
-  - astropy.io.votable is used for parsing VOTable files from the SVO service
