+++ pyphot/typing.py
@@ -1,0 +1,15 @@
+"""Module containing type aliases for pyphot."""
+
+from .unit_adapters.typing import (
+    UnitAdapterType,
+    QuantityType,
+    DimensionalityError,
+    UndefinedUnitError,
+)
+
+__all__ = [
+    "UnitAdapterType",
+    "QuantityType",
+    "DimensionalityError",
+    "UndefinedUnitError",
+]
