### 2025-11-06 ddb01ac
add get_library
