+++ pyphot/__init__.py
@@ -1,12 +1,12 @@
 from . import typing
 from . import legacy
-from .version import __VERSION__
 
-from .phot import Filter
-from .libraries import Library, HDF_Library, Ascii_Library
+from .libraries import Library, HDF_Library, Ascii_Library, get_library
 from .licks import LickLibrary, LickIndex, reduce_resolution
-from .vega import Vega
+from .phot import Filter
 from .sun import Sun
+from .vega import Vega
+from .version import __VERSION__
 
 __all__ = [
     "Ascii_Library",
@@ -18,6 +18,7 @@ __all__ = [
     "Sun",
     "Vega",
     "__VERSION__",
+    "get_library",
     "legacy",
     "reduce_resolution",
     "typing",
