### 2025-11-06 7074197
ruffed
