+++ pyphot/io/hdf.py
@@ -1,5 +1,9 @@
-from typing import Optional
+"""Read and write HDF5 files with pytables preserving metadata (tables, https://www.pytables.org/)"""
+
+from typing import Optional, Literal, Union
 import pandas as pd
+from os import PathLike
+
 from .header import HeaderInfo
 
 import tables
@@ -92,3 +96,142 @@ def from_hdf5(
 
         return pd.DataFrame(data), hdr
     raise ValueError("Something went wrong without much information from pytables")
+
+
+def to_hdf(
+    df: pd.DataFrame,
+    filename: Union[str, tables.File, PathLike],
+    *,
+    tablename: Optional[str] = None,
+    header_info: Optional[HeaderInfo] = None,
+    mode: Literal["r", "w", "a", "r+"] = "w",
+    append: bool = False,
+    **kwargs,
+) -> None:
+    """
+    Write a pandas DataFrame to an HDF5 file.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        The DataFrame to write.
+    filename : str or tables.File or PathLike
+        The filename or open HDF5 file to write to.
+    tablename : str, optional
+        The name of the table to write to.
+    header_info : HeaderInfo, optional
+        The header information to write. Default is to use from df.attrs
+    mode : {'r', 'w', 'a', 'r+'}, default 'w'
+        The mode to open the file in.
+    append : bool, default False
+        Whether to append data to an existing file.
+    **kwargs
+        Additional keyword arguments to pass to tables.open_file.
+
+    Raises
+    ------
+    Exception
+        If the HDF backend does not implement stream.
+    tables.FileModeError
+        If the file is already opened in a different mode.
+    ValueError
+        If something went wrong without much information from pytables.
+    """
+    if hasattr(filename, "read"):
+        raise Exception("HDF backend does not implement stream")
+
+    # ensure mode is valid for appending data
+    mode = "a" if append is True else mode
+
+    # open output file, or if provided tables.File, check it's in the correct mode
+    if isinstance(filename, tables.File):
+        if (filename.mode != mode) & (mode != "r"):
+            raise tables.FileModeError("The file is already opened in a different mode")
+        hd5 = filename
+    else:
+        hd5 = tables.open_file(str(filename), mode=mode)
+
+    if header_info is None:
+        # attempt to get it from the dataframe attributes
+        header_info = HeaderInfo(
+            header={
+                k: v
+                for k, v in df.attrs.items()
+                if k not in ["aliases", "units", "comments"]
+            },
+            alias=df.attrs.get("aliases", {}),
+            units=df.attrs.get("units", {}),
+            comments=df.attrs.get("comments", {}),
+        )
+
+    # check table name and path
+    tablename_path = (
+        tablename
+        or df.name
+        or header_info.header.get("name", None)
+        or header_info.header.get("NAME", None)
+    )
+    if tablename_path in ("", None, "Noname", "None"):
+        tablename_path = "/data"
+    if not tablename_path.startswith("/"):
+        tablename_path = "/" + tablename_path
+
+    if append:
+        try:
+            tab = hd5.get_node(tablename_path)
+            if not isinstance(tab, tables.Table):
+                raise TypeError("Node is not a table")
+            if tab.description is None:
+                raise ValueError("Table description is missing")
+            dtypes = tab.description._v_dtype  # pyright: ignore / it is there
+            data = df.to_records(index=False, column_dtypes=dtypes)
+            tab.append(data)
+            tab.flush()
+        except tables.NoSuchNodeError:
+            print(
+                f"Warning: Table {tablename_path} does not exist. A new table will be created."
+            )
[…]
