### 2025-11-06 5741086
adds from_file to auto try readers
