+++ .github/workflows/unittests.yml
@@ -26,8 +26,7 @@ jobs:
       - name: Running tests
         run: |
           python3 -m pytest --cov pyphot
+          DATA=$(python3 -mcoverage report)
           echo "Coverage report"
           echo "# Coverage Report" >> $GITHUB_STEP_SUMMARY
-          echo "```" >> $GITHUB_STEP_SUMMARY
-          coverage report >> $GITHUB_STEP_SUMMARY
-          echo "```\n" >> $GITHUB_STEP_SUMMARY
+          echo "```$DATA```" >> $GITHUB_STEP_SUMMARY
