### 2025-11-06 7afc362
workflow summary output
