+++ .github/workflows/unittests.yml
@@ -26,7 +26,9 @@ jobs:
       - name: Running tests
         run: |
           python3 -m pytest --cov pyphot
-          DATA=$(python3 -mcoverage report)
+          DATA=$(python3 -m coverage report)
           echo "Coverage report"
           echo "# Coverage Report" >> $GITHUB_STEP_SUMMARY
-          echo "```$DATA```" >> $GITHUB_STEP_SUMMARY
+          echo "```" >> $GITHUB_STEP_SUMMARY
+          echo "$DATA" >> $GITHUB_STEP_SUMMARY
+          echo "```" >> $GITHUB_STEP_SUMMARY
