### 2025-11-06 a9d2a4a
workflow summary
