+++ .github/workflows/unittests.yml
@@ -27,7 +27,7 @@ jobs:
         run: |
           python3 -m pytest --cov pyphot
           echo "# Coverage Report" >> $GITHUB_STEP_SUMMARY
-          echo "```" >> $GITHUB_STEP_SUMMARY
+          echo '```' >> $GITHUB_STEP_SUMMARYY
           echo "--- Coverage Report Output (Console & Summary) ---"
           python3 -m coverage report | tee -a /dev/stderr | cat >> $GITHUB_STEP_SUMMARY
-          echo "```" >> $GITHUB_STEP_SUMMARY
+          echo '```' >> $GITHUB_STEP_SUMMARYY
