+++ doc/index.rst
@@ -6,81 +6,73 @@
 PYPHOT -- A tool for computing photometry from spectra
 ======================================================
 
-This is a set of tools to compute synthetic photometry in a simple way, suitable for
-integration in larger projects.
+This is a set of tools to compute synthetic photometry in a simple way, suitable
+for integration in larger projects.
 
-The inputs are response functions for the desired photometric passbands, and stellar spectra. The modules are flexible to handle units in the wavelength definition through 
-a simplified version of `pint` (link)
+The inputs are response functions for the desired photometric passbands, and
+stellar spectra.
 
 Filters are represented individually by a `Filter` object. Collections of
 filters are handled with a `Library`. We provide an internal library that
-contains a signitificant amount of common filters. Note that we use the word "filter" 
-in a generic sense: in many cases it corresponds in fact to a total throughput, that includes
-the actual filter transmission, the transmission of the optics, the wavelength-dependent efficiency
-of the detector, and (for ground-based photometric systems) the transmission of the 
-atmosphere.
+contains a signitificant amount of common filters. Note that we use the word
+"filter" in a generic sense: in many cases it corresponds in fact to a total
+throughput, that includes the actual filter transmission, the transmission of
+the optics, the wavelength-dependent efficiency of the detector, and (for
+ground-based photometric systems) the transmission of the atmosphere.
 
-Each filter is minimally defined by a `wavelength`, a `throughput`, and a detector
-type that may be either `energy` or `photon` (default). Many properties such as central 
-or pivot wavelengths are computed internally. 
+Each filter is minimally defined by a `wavelength`, a `throughput`, and a
+detector type that may be either `energy` or `photon` (default). Many properties
+such as central or pivot wavelengths are computed internally.
 
-Units are provided for the wavelengths of the filters, zero points in
-multiple units are also accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The
-default detector is assumed to be a photon-counting device, but energy-sensitive
+Units are provided for the wavelengths of the filters, zero points in multiple
+units are also accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The default
+detector is assumed to be a photon-counting device, but energy-sensitive
 detectors are also handled for the computations.
 
 .. note::
 
-        All provided filters have defined units and detector type
-        that are used transparently throughout the package.
+    **New in v2.0.0**
 
-.. image:: https://mybinder.org/badge.svg 
+    Pyphot handles units from `Astropy Units <https://docs.astropy.org/en/stable/units/index.html>`_ (as default), `Pint <https://pint.readthedocs.io/en/stable/>`_, and `EzUnits`_ (:mod:`~pyphot.unit_adapters.ezunits`).
+    You can switch between them using the :func:`pyphot.config.set_units_backend` function.
+
+    .. code-block:: python
+
+        import pyphot.config
+        pyphot.config.set_units_backend('astropy')
+        pyphot.config.set_units_backend('pint')
+        pyphot.config.set_units_backend('ezunits')
+
+.. image:: https://mybinder.org/badge.svg
   :target: https://mybinder.org/v2/gh/mfouesneau/pyphot/master?filepath=examples%2FQuickStart.ipynb
- 
+
 .. image:: https://img.shields.io/badge/render%20on-nbviewer-orange.svg
[…]
