+++ doc/pyphot.astropy.rst
@@ -1,45 +1,0 @@
-pyphot.astropy package
-======================
-
-Submodules
-----------
-
-pyphot.astropy.config module
-----------------------------
-
-.. automodule:: pyphot.astropy.config
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyphot.astropy.sandbox module
------------------------------
-
-.. automodule:: pyphot.astropy.sandbox
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyphot.astropy.sun module
--------------------------
-
-.. automodule:: pyphot.astropy.sun
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pyphot.astropy.vega module
---------------------------
-
-.. automodule:: pyphot.astropy.vega
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pyphot.astropy
-   :members:
-   :undoc-members:
-   :show-inheritance:
