+++ doc/pyphot.io.rst
@@ -1,0 +1,53 @@
+pyphot.io package
+=================
+
+Submodules
+----------
+
+pyphot.io.ascii module
+----------------------
+
+.. automodule:: pyphot.io.ascii
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.fits module
+---------------------
+
+.. automodule:: pyphot.io.fits
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.hdf module
+--------------------
+
+.. automodule:: pyphot.io.hdf
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.header module
+-----------------------
+
+.. automodule:: pyphot.io.header
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.votable module
+------------------------
+
+.. automodule:: pyphot.io.votable
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: pyphot.io
+   :members:
+   :show-inheritance:
+   :undoc-members:
