+++ doc/pyphot.unit_adapters.ezunits.rst
@@ -1,0 +1,21 @@
+pyphot.unit\_adapters.ezunits package
+=====================================
+
+Submodules
+----------
+
+pyphot.unit\_adapters.ezunits.pint module
+-----------------------------------------
+
+.. automodule:: pyphot.unit_adapters.ezunits.pint
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Module contents
+---------------
+
+.. automodule:: pyphot.unit_adapters.ezunits
+   :members:
+   :show-inheritance:
+   :undoc-members:

[…]
