### 2025-11-11 1b09c52
add major additiional information
