+++ doc/licks.rst
@@ -1,40 +1,144 @@
 Extention to Lick indices
 =========================
 
-We also include functions to compute lick indices and provide a series of
-commonly use ones.
+We also include functions to compute lick indices and provide a series of commonly use ones.
 
 Lick index
 ~~~~~~~~~~
 
-The Lick system of spectral line indices is one of the most commonly used
-methods of determining ages and metallicities of unresolved (integrated light)
-stellar populations.
+The Lick system of spectral line indices is one of the most commonly used methods of determining ages and metallicities of unresolved (integrated light) stellar populations.
 
-The calibration of the Lick / IDS system is complicated because the original
-Lick spectra were not flux calibrated, so there are usually systematic effects
-due to differences in continuum shape.  Proper calibration involves observing
-many of the original Lick/IDS standard stars and deriving offsets to the
-standard system.
+The calibration of the Lick / IDS system is complicated because the original Lick spectra were not flux calibrated, so there are usually systematic effects due to differences in continuum shape.  Proper calibration involves observing many of the original Lick/IDS standard stars and deriving offsets to the standard system.
 
 In `Vazdekis et al. (2010) <https://ui.adsabs.harvard.edu/abs/2010MNRAS.404.1639V>`_, they propose a new Line Index System, hereafter
-LIS, with three new spectral resolutions at which to measure the Lick
-indices. Note that this new system should not be restricted to the Lick set
-of indices in a flux calibrated system. In fact, LIS can be used for any
-index in the literature (e.g., for the Rose (1984) indices), including
-newly defined indices (e.g., Cervantes & Vazdekis 2009).
+LIS, with three new spectral resolutions at which to measure the Lick indices. Note that this new system should not be restricted to the Lick set of indices in a flux calibrated system. In fact, LIS can be used for any index in the literature (e.g., for the `Rose (1984) indices <https://ui.adsabs.harvard.edu/abs/1985AJ.....90.1927R>`_), including newly defined indices (e.g., `Cervantes & Vazdekis 2009 <https://ui.adsabs.harvard.edu/abs/2009MNRAS.392..691C>`_).
 
 
-The LIS system is defined for 3 different spectral resolutions which are
-best suited for the following astrophysical cases:
+The LIS system is defined for 3 different spectral resolutions which are best suited for the following astrophysical cases:
 
 * LIS-5.0AA: globular clusters
 * LIS-8.4AA: low and intermediate-mass galaxies
 * LIS-14.0AA: massive galaxies
 
-Conversions to transform the data from the Lick/IDS system to LIS can be found
-in Johansson, Thomas & Maraston (2010), which provides a discussion of indices
+Conversions to transform the data from the Lick/IDS system to LIS can be found in `Johansson, Thomas & Maraston (2010) <https://ui.adsabs.harvard.edu/abs/2010MNRAS.404.1639V>`_, which provides a discussion of indices
 and the information content of them.
+
[…]
