+++ doc/index.rst
@@ -6,8 +6,27 @@
 PYPHOT -- A tool for computing photometry from spectra
 ======================================================
 
+.. image:: https://img.shields.io/pypi/v/pyphot.svg
+    :target: https://pypi.org/project/pyphot/
+
+.. image:: https://zenodo.org/badge/70060728.svg
+    :target: https://zenodo.org/badge/latestdoi/70060728
+
+.. image:: https://static.pepy.tech/badge/pyphot
+    :target: https://pepy.tech/project/pyphot
+
+.. image:: https://static.pepy.tech/badge/pyphot/month
+    :target: https://pepy.tech/project/pyphot
+
+.. image:: https://img.shields.io/badge/python-3.9,_3.10,_3.11,_3.12,_3.13-blue.svg
+
 This is a set of tools to compute synthetic photometry in a simple way, suitable
 for integration in larger projects.
+
+.. important::
+   New major release (v2.0.0) which introduces new features and **breaking changes**.
+
+   See :doc:`What's new <whats_new>`.
 
 The inputs are response functions for the desired photometric passbands, and
 stellar spectra.
@@ -31,9 +50,7 @@ detectors are also handled for the computations.
 
 .. note::
 
-    **New in v2.0.0**
-
-    Pyphot handles units from `Astropy Units <https://docs.astropy.org/en/stable/units/index.html>`_ (as default), `Pint <https://pint.readthedocs.io/en/stable/>`_, and `EzUnits`_ (:mod:`~pyphot.unit_adapters.ezunits`).
+    Pyphot handles units from `Astropy Units <https://docs.astropy.org/en/stable/units/index.html>`_ (as default), `Pint <https://pint.readthedocs.io/en/stable/>`_, and `EzUnits` (:mod:`~pyphot.unit_adapters.ezunits`).
     You can switch between them using the :func:`pyphot.config.set_units_backend` function.
 
     .. code-block:: python
@@ -43,11 +60,6 @@ detectors are also handled for the computations.
         pyphot.config.set_units_backend('pint')
         pyphot.config.set_units_backend('ezunits')
 
-.. image:: https://mybinder.org/badge.svg
-  :target: https://mybinder.org/v2/gh/mfouesneau/pyphot/master?filepath=examples%2FQuickStart.ipynb
-
-.. image:: https://img.shields.io/badge/render%20on-nbviewer-orange.svg
-  :target: https://nbviewer.jupyter.org/github/mfouesneau/pyphot/tree/master/examples/
 
 
 Package main content
@@ -74,14 +86,16 @@ Additionally, and for convenience
 
 * :func:`pyphot.svo.get_pyphot_filter` provides an interface to the `SVO Filter Profile Service <http://svo2.cab.inta-csic.es/theory/fps/>`_ to download filters directly from the SVO database.
 
-Contents
----------
+Documentation Contents
+----------------------
 
 .. toctree::
    :maxdepth: 1
 
    Home <index>
-   more_examples
+   What's New <whats_new>
+   QuickStart
+   examples
    photometry
    libcontent
    licks
@@ -93,9 +107,7 @@ Installation
 ~~~~~~~~~~~~
 Pyphot is available on `PyPI <https://pypi.org/project/pyphot/>`_ and can be installed using any common package manager.
 
-
-
-* Using pip: Use the `--user` option if you don't have permissions to install libraries
+* Using pip: (Use the `--user` option if you don't have permissions to install libraries)
 
 .. image:: https://img.shields.io/pypi/v/pyphot.svg
     :target: https://pypi.org/project/pyphot/
@@ -104,94 +116,13 @@ Pyphot is available on `PyPI <https://pypi.org/project/pyphot/>`_ and can be ins
[…]
