### 2025-11-11 2cb6bdb
Fix scripts to v2.0 api
