+++ pyphot/io/fits.py
@@ -236,8 +236,7 @@ def to_fits(
         Output verification option.  Must be one of ``"fix"``, ``"silentfix"``,
         ``"ignore"``, ``"warn"``, or ``"exception"``.  May also be any
         combination of ``"fix"`` or ``"silentfix"`` with ``"+ignore"``,
-        ``+warn``, or ``+exception" (e.g. ``"fix+warn"``).  See :ref:`verify`
-        for more info.
+        ``+warn``, or ``+exception" (e.g. ``"fix+warn"``).
     checksum : bool, optional
         If `True`, adds both ``DATASUM`` and ``CHECKSUM`` cards to the
         headers of all HDU's written to the file
