+++ pyphot/io/ascii.py
@@ -328,28 +328,28 @@ def to_csv(
         A string representing the encoding to use in the output file,
         defaults to 'utf-8'. `encoding` is not supported if `path_or_buf`
         is a non-binary file object.
-        compression : str or dict, default 'infer'
-            For on-the-fly compression of the output data. If 'infer' and 'path_or_buf' is
-            path-like, then detect compression from the following extensions: '.gz',
-            '.bz2', '.zip', '.xz', '.zst', '.tar', '.tar.gz', '.tar.xz' or '.tar.bz2'
-            (otherwise no compression).
-            Set to ``None`` for no compression.
-            Can also be a dict with key ``'method'`` set
-            to one of {``'zip'``, ``'gzip'``, ``'bz2'``, ``'zstd'``, ``'xz'``, ``'tar'``} and
-            other key-value pairs are forwarded to
-            ``zipfile.ZipFile``, ``gzip.GzipFile``,
-            ``bz2.BZ2File``, ``zstandard.ZstdCompressor``, ``lzma.LZMAFile`` or
-            ``tarfile.TarFile``, respectively.
-            As an example, the following could be passed for faster compression and to create
-            a reproducible gzip archive:
-            ``compression={'method': 'gzip', 'compresslevel': 1, 'mtime': 1}``.
+    compression : str or dict, default 'infer'
+        For on-the-fly compression of the output data. If 'infer' and 'path_or_buf' is
+        path-like, then detect compression from the following extensions: '.gz',
+        '.bz2', '.zip', '.xz', '.zst', '.tar', '.tar.gz', '.tar.xz' or '.tar.bz2'
+        (otherwise no compression).
+        Set to ``None`` for no compression.
+        Can also be a dict with key ``'method'`` set
+        to one of {``'zip'``, ``'gzip'``, ``'bz2'``, ``'zstd'``, ``'xz'``, ``'tar'``} and
+        other key-value pairs are forwarded to
+        ``zipfile.ZipFile``, ``gzip.GzipFile``,
+        ``bz2.BZ2File``, ``zstandard.ZstdCompressor``, ``lzma.LZMAFile`` or
+        ``tarfile.TarFile``, respectively.
+        As an example, the following could be passed for faster compression and to create
+        a reproducible gzip archive:
+        ``compression={'method': 'gzip', 'compresslevel': 1, 'mtime': 1}``.
 
-            May be a dict with key 'method' as compression mode
-            and other entries as additional compression options if
-            compression mode is 'zip'.
+        May be a dict with key 'method' as compression mode
+        and other entries as additional compression options if
+        compression mode is 'zip'.
 
-            Passing compression options as keys in dict is
-            supported for compression modes 'gzip', 'bz2', 'zstd', and 'zip'.
+        Passing compression options as keys in dict is
+        supported for compression modes 'gzip', 'bz2', 'zstd', and 'zip'.
     quoting : optional constant from csv module
         Defaults to csv.QUOTE_MINIMAL. If you have set a `float_format`
         then floats are converted to strings and thus csv.QUOTE_NONNUMERIC
