+++ doc/photometry.rst
@@ -5,7 +5,7 @@ It is sometimes not obvious that there are important differences between photome
 
 In this section, we review the important details for computing the luminosity and the magnitude of a star through a photometric passband. We do not discuss calibration, which in principle is covered by instrument documentations.
 
-Synthetic photometry uses transmission curves that typically include a variety of wavelength-dependent components (a filter transmission, the response of the optical elements of the telescope and instrument, the detector sensitivity, sometimes a telluric absorption component, ...). Transmission curves are usually named after the associated filter, and it is up to the user to verify that the other components are also included. In this documentation, we use the terms filter throughput, transmission curve, response function, or simply filter, interchangeably, to refer to a transmission curve, generally leaving the verification of the nature of these transmission curves to the user. 
+Synthetic photometry uses transmission curves that typically include a variety of wavelength-dependent components (a filter transmission, the response of the optical elements of the telescope and instrument, the detector sensitivity, sometimes a telluric absorption component, ...). Transmission curves are usually named after the associated filter, and it is up to the user to verify that the other components are also included. In this documentation, we use the terms filter throughput, transmission curve, response function, or simply filter, interchangeably, to refer to a transmission curve, generally leaving the verification of the nature of these transmission curves to the user.
 
 Most modern detectors count photons, rather than cumulating the energies of these photons.  If we consider a filter throughput (a.k.a, transmission curve, or response function) defined in wavelength by the dimensionless function :math:`T(\lambda)`, this function tells you what fraction of the arriving photons at wavelength :math:`\lambda` actually get detected (on average).  Therefore, the total number of photons, per unit time per unit collecting area, expected to be detected in this filter is
 
@@ -17,7 +17,7 @@ Most modern detectors count photons, rather than cumulating the energies of thes
 
 where :math:`f_\lambda` is the wavelength dependent flux density of an object given in energy per unit time, per unit collecting area, and per unit wavelength.
 
-Consequently, interpreting :math:`\lambda T(\lambda)` as a distribution leads to the statistical mean of the flux density, :math:`\overline{f_\lambda}` 
+Consequently, interpreting :math:`\lambda T(\lambda)` as a distribution leads to the statistical mean of the flux density, :math:`\overline{f_\lambda}`
 
 .. math::
 
@@ -29,17 +29,17 @@ Note that although this is formally a weighted mean of a flux density, with weig
 
 .. code-block:: python
 
-        # computing the mean flux of a spectrum 
-        # single_spectrum: np.array of shape (n_lambda,)
-        flux = lib['hst_wfc3_f110w'].get_flux(λ, single_spectrum)
-        # λ may have units, otherwise assuming consistent definitions.
+    # computing the mean flux of a spectrum
+    # single_spectrum: np.array of shape (n_lambda,)
+    flux = lib['hst_wfc3_f110w'].get_flux(λ, single_spectrum)
+    # λ may have units, otherwise assuming consistent definitions.
 
-        # computing the mean flux of many spectra
[…]
