### 2025-11-11 e4729ea
Update docstrings for API documentation
