+++ pyphot/unit_adapters/__init__.py
@@ -1,5 +1,54 @@
 """
-Checks which backends are available and provides a global typing hint for the units adapters.
+This module provides a framework for using different unit packages such as astropy, pint, and others.
+It provides a unified interface for working with units across different packages through adapters.
+
+Registered Adapters
+~~~~~~~~~~~~~~~~~~~
+
+We provide the following adapters:
+
+- :class:`~pyphot.unit_adapters.AstropyAdapter`: Adapter for the astropy unit system.
+- :class:`~pyphot.unit_adapters.PintAdapter`: Adapter for the pint unit system.
+- :class:`~pyphot.unit_adapters.EzUnitsAdapter`: Adapter for the legacy pyphot unit system based on pint (v0.1.0).
+
+Note: The :class:`~pyphot.unit_adapters.EzUnitsAdapter` is provided as a fallback for older codes and in case other unit packages are missing.
+
+API Overview
+~~~~~~~~~~~~
+
+All adapters must implement the :class:`~pyphot.unit_adapters.units_adapter.UnitsAdapter` interface, which defines the following methods:
+
+- :meth:`~pyphot.unit_adapters.units_adapter.UnitsAdapter.U()`: queries the unit registry.
+- :meth:`~pyphot.unit_adapters.units_adapter.UnitsAdapter.Q()`: make sure the returned object is a quantity, i.e. a value with unit (e.g. astropy differs between `U("kg")` and `Q("kg") = 1. * U("kg")`)
+
+For source typing reasons, adapters also override a few methods:
+
+- :meth:`~pyphot.unit_adapters.units_adapter.UnitsAdapter.has_unit`: checks if the object has a unit.
+- :meth:`~pyphot.unit_adapters.units_adapter.UnitsAdapter.val_in_unit`: returns the value of the quantity in the specified unit or forces a default unit if not specified.
+
+Decorator to impose units on arguments and keyword arguments
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+For convenience, adapters provide (by inheritance) a `decorate` decorator which can help defining functions with default units on arguments and keyword arguments.
+
+For a more general decorator, use :func:`enforce_default_units` which will assumes the defined global adapter, if not a specific one provided.
+
+.. note::
+    Both decoration methods will update the docstring of the decorated function to include information about the imposed units.
+
+Typing
+~~~~~~
+
+Hint typing information is provided at the module level: :class:`~pyphot.unit_adapters.UnitAdapterType`, :class:`~pyphot.unit_adapters.QuantityType`, which is a union of :class:`~pyphot.unit_adapters.units_adapter.UnitsAdapter` and varied :class:`~pyphot.unit_adapters.Quantity` definitions.
+In addition, each adapter provides specific types in their `typing.Quantity` attribute.
+
+Exceptions
+~~~~~~~~~~
+
+For convenience adapters also provide a list of potential exception types. At the module level
+:class:`~pyphot.unit_adapters.UnitsAdapter.UndefinedUnitError`, and :class:`~pyphot.unit_adapters.UnitsAdapter.DimensionalityError` which applies to any backend units.
+
+When needed :class:`~pyphot.unit_adapters.UnitsAdapter.typing.UndefinedUnitError` and :class:`~pyphot.unit_adapters.UnitsAdapter.typing.DimensionalityError` can be used to catch specific errors related to unit conversion or registration by the unit package.
 """
 
 from typing import Any
