### 2025-11-11 3552e55
add skip copying documentation of submodules from __init__ calls.
