### 2025-11-11 c95ac9e
remove docstring tests
