### 2025-11-11 cf1c001
RC pyphot v2.0 (#90)
RC pyphot v2.0

## Breaking changes

- major refactoring of the codebase: remove code duplication due to the split astropy vs. pint
    - :code:`import pyphot.astropy as pyphot` replaced by :code:`import pyphot` (default backend is 'astropy')
    - set units with :code:`pyphot.config.set_units_backend('astropy')` or :code:`pyphot.config.set_units_backend('pint')` (default is 'astropy')
- adapters to interface transparently with varied unit frameworks (e.g., astropy.units, pint)
  - :func:`pyphot.config.set_units_backend` to switch between different unit frameworks
  - defa
[…]
