+++ unittests/test_quick.py
@@ -1,314 +1,0 @@
-import unittest
-import sys
-sys.path.append('../')
-
-class TestQuick(unittest.TestCase):
-    def test_dependencies(self):
-        from pyphot import sandbox as pyphot
-
-    def test_library_irac(self):
-        from pyphot import sandbox as pyphot
-        # get the internal default library of passbands filters
-        lib = pyphot.get_library()
-        print("Library contains: ", len(lib), " filters")
-        # find all filter names that relates to IRAC
-        # and print some info
-        f = lib.find('irac')
-        for name in f:
-            lib[name].info(show_zeropoints=True)
-
-    def test_vega(self):
-        from pyphot.vega import Vega
-        vega = Vega()
-
-    def test_convert_mags(self):
-        import numpy as np
-        from pyphot.vega import Vega
-        from pyphot import sandbox as pyphot
-        # get the internal default library of passbands filters
-        lib = pyphot.get_library()
-        vega = Vega()
-        f = lib['HST_WFC3_F110W']
-        # compute the integrated flux through the filter f
-        # note that it work on many spectra at once
-        fluxes = f.get_flux(vega.wavelength, vega.flux, axis=-1)
-
-        # Note that fluxes is now with units of erg/s/cm2/AA
-        # pyphot gives Vega in flam and can convert between flux density units.
-        print(fluxes, vega.wavelength, vega.flux)
-
-        # convert to vega magnitudes
-        mags = -2.5 * np.log10(fluxes.magnitude) - f.Vega_zero_mag
-        print("Vega magnitude of Vega in {0:s} is : {1:f} mag".format(f.name, mags))
-        mags = -2.5 * np.log10(fluxes.magnitude) - f.AB_zero_mag
-        print("AB magnitude of Vega in {0:s} is : {1:f} mag".format(f.name, mags))
-        mags = -2.5 * np.log10(fluxes.magnitude) - f.ST_zero_mag
-        print("ST magnitude of Vega in {0:s} is : {1:f} mag".format(f.name, mags))
-
-    def test_load_all_filters(self):
-        from pyphot import sandbox as pyphot
-        # get the internal default library of passbands filters
-        lib = pyphot.get_library()
-        lib.load_all_filters()
-
-    def test_all_filters_in_library(self):
-        from pyphot import sandbox as pyphot
-
-        # define header and table format (as csv)
-        hdr = ("name", "detector type", "wavelength units",
-            "central wavelength", "pivot wavelength", "effective wavelength",
-            "Vega mag", "Vega flux", "Vega Jy",
-            "AB mag", "AB flux", "AB Jy",
-            "ST mag", "ST flux", "ST Jy")
-        fmt = "{0:s},{1:s},{2:s},{3:.3f},{4:.3f},{5:.3f},{6:.5f},{7:.5g},{8:.5g},{9:.5f},{10:.5g},{11:.5g},{12:.5f},{13:.5g},{14:.5g}\n"
-
-        l = pyphot.get_library()
-
-        with open('table.csv', 'w') as output:
-            output.write(','.join(hdr) + '\n')
-
-            for k in sorted(l.content):
-                fk = l[k]
-                rec = (fk.name, fk.dtype, fk.wavelength_unit,
-                    fk.cl.magnitude, fk.lpivot.magnitude, fk.leff.magnitude,
-                    fk.Vega_zero_mag, fk.Vega_zero_flux.magnitude, fk.Vega_zero_Jy.magnitude,
[…]
