### 2025-11-11 b498743
Update release date for major refactoring entry
