### 2025-11-12 d623570
add pre-commit stage
