+++ pyphot/unit_adapters/ezunits/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
 pint
 ~~~~
