+++ _sources/index.rst
@@ -1,0 +1,264 @@
+.. pyphot documentation master file, created by
+   sphinx-quickstart on Wed Oct  5 11:25:47 2016.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+PYPHOT -- A tool for computing photometry from spectra
+======================================================
+
+.. image:: https://img.shields.io/pypi/v/pyphot.svg
+    :target: https://pypi.org/project/pyphot/
+
+.. image:: https://zenodo.org/badge/70060728.svg
+    :target: https://zenodo.org/badge/latestdoi/70060728
+
+.. image:: https://static.pepy.tech/badge/pyphot
+    :target: https://pepy.tech/project/pyphot
+
+.. image:: https://static.pepy.tech/badge/pyphot/month
+    :target: https://pepy.tech/project/pyphot
+
+.. image:: https://img.shields.io/badge/python-3.9,_3.10,_3.11,_3.12,_3.13-blue.svg
+
+This is a set of tools to compute synthetic photometry in a simple way, suitable
+for integration in larger projects.
+
+.. important::
+   New major release (v2.0.0) which introduces new features and **breaking changes**.
+
+   See :doc:`What's new <whats_new>`.
+
+.. info::
+   Support for python 3.14 is currently blocked by `pytables` not supporting it yet.
+
+
+The inputs are response functions for the desired photometric passbands, and
+stellar spectra.
+
+Filters are represented individually by a `Filter` object. Collections of
+filters are handled with a `Library`. We provide an internal library that
+contains a signitificant amount of common filters. Note that we use the word
+"filter" in a generic sense: in many cases it corresponds in fact to a total
+throughput, that includes the actual filter transmission, the transmission of
+the optics, the wavelength-dependent efficiency of the detector, and (for
+ground-based photometric systems) the transmission of the atmosphere.
+
+Each filter is minimally defined by a `wavelength`, a `throughput`, and a
+detector type that may be either `energy` or `photon` (default). Many properties
+such as central or pivot wavelengths are computed internally.
+
+Units are provided for the wavelengths of the filters, zero points in multiple
+units are also accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The default
+detector is assumed to be a photon-counting device, but energy-sensitive
+detectors are also handled for the computations.
+
+.. note::
+
+    Pyphot handles units from `Astropy Units <https://docs.astropy.org/en/stable/units/index.html>`_ (as default), `Pint <https://pint.readthedocs.io/en/stable/>`_, and `EzUnits` (:mod:`~pyphot.unit_adapters.ezunits`).
+    You can switch between them using the :func:`pyphot.config.set_units_backend` function.
+
+    .. code-block:: python
+
+        import pyphot.config
+        pyphot.config.set_units_backend('astropy')
+        pyphot.config.set_units_backend('pint')
+        pyphot.config.set_units_backend('ezunits')
+
+
+
+Package main content
+~~~~~~~~~~~~~~~~~~~~
+
+This package is mostly organized around 2 main classes:
+
+* :class:`~pyphot.phot.Filter` that handles filter definitions and manipulations.
+
+* :class:`~pyphot.libraries.Library` that handles a collection of filters in a few formats.
+
+
+Additionally, and for convenience
+
+* the library class is derived into an ascii file reader
+  :class:`~pyphot.libraries.Ascii_Library`, and a single hdf file reader
+  :class:`~pyphot.libraries.HDF_Library`.
+
[…]
