### 2025-12-04 mfouesneau/pyphot#91 @WillJRoper
Minor README updates required
There are some minor issues in the README:
- The pint link is missing with "pint (link)" instead of the link.
- You are missing the EzUnits package which is included in the docs.

Very very minor:
- Missing capital in "full documentation".
- Might be intentional but line break after "Note" looks unintentional.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8814#issuecomment-3451697350
