### 2025-12-04 0aa2f4a
Refine language and fix grammar in README.rst
Updated wording for clarity and corrected minor grammatical issues throughout the README.
#91
