+++ README.rst
@@ -15,25 +15,15 @@ pyphot -- A tool for computing photometry from spectra
 
 .. image:: https://img.shields.io/badge/python-3.9,_3.10,_3.11,_3.12,_3.13-blue.svg
 
-This is a set of tools to compute synthetic photometry in a simple way, ideal to
-integrate in larger projects.
+This is a set of tools for computing synthetic photometry in a simple way, ideal for integration into larger projects.
 
-full documentation at: http://mfouesneau.github.io/pyphot/
+Full documentation at: http://mfouesneau.github.io/pyphot/
 
-The inputs are photonic or energetic response functions for the desired
-photometric bands and stellar spectra. The modules are flexible to handle units
-in the wavelength definition through a simplified version of `pint` (link) and `astropy.units`
+The inputs are photonic or energetic response functions for the desired photometric bands and stellar spectra. The modules are flexible to handle units in the wavelength definition through a simplified version of `pint <https://pint.readthedocs.io/en/stable/>`_  and `astropy.units <https://docs.astropy.org/en/stable/units/index.html>`_
 
-Filters are represented individually by a `Filter` object. Collections of
-filters are handled with a `Library`. We provide an internal library that
-contains a signitificant amount of common filters.
+Filters are represented individually by a `Filter` object. Collections of filters are handled with a `Library`. We provide an internal library that contains a significant number of common filters.
 
-Each filter is minimally defined by a `wavelength` and `throughput`. Many
-properties such as central of pivot wavelength are computed internally. When
-units are provided for the wavelength, zero points in multiple units are also
-accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The default detector type is
-assumed to be photonic, but energetic detectors are also handled for the
-computations.
+Each filter is minimally defined by a `wavelength` and `throughput`. Many properties, such as the central or pivot wavelength, are computed internally. When units are provided for the wavelength, zero points in multiple units are also accessible (AB, Vega magnitude, Jy, erg/s/cm2/AA). The default detector type is assumed to be photonic, but energetic detectors are also handled for the computations.
 
 .. image:: https://mybinder.org/badge.svg
   :target: https://mybinder.org/v2/gh/mfouesneau/pyphot/master?filepath=examples%2FQuickStart.ipynb
@@ -41,30 +31,13 @@ computations.
 .. image:: https://img.shields.io/badge/render%20on-nbviewer-orange.svg
   :target: https://nbviewer.jupyter.org/github/mfouesneau/pyphot/tree/master/examples/
 
-
-
-.. note::
-    `cphot <https://github.com/mfouesneau/cphot>`_ is a spin-off project that provides a C++ version of pyphot
-
 What's new?
 -----------
-* [November 11, 2025] major refactoring, remove code duplication, added type hints (v2.0.0) -- See `release notes <https://mfouesneau.github.io/pyphot/whats_new.html>`_
-* [July 2, 2025] Added Vega flavors to the Vega class, allowing users to select different Vega reference spectra for photometric calculations. (v1.7.0)
-* [January 21, 2025] Moved installation to use `pyproject.toml`; Updated documentation; Updates for deprecated calls. (v1.6.0)
-* [January 21, 2025] Support for Python 3.13 (v1.5.0);
-* [June 26, 2024] Dropping support for Python <= 3.8 (due to HDF5 modules). Minor updates for Scipy 1.14.0 and Numpy 2.0
-* [November 22, 2021] new filters, SVO interface, automated tests and documentation.
[…]
