### 2025-12-04 cada965
Fix installation instructions
Corrected typos and improved clarity in documentation.
#88
