### 2025-12-04 mfouesneau/pyphot#92 @WillJRoper
Some modules are missing doc strings or doc string content
There is some inconsistency in whether doc strings are include and are complete across the codebase. Flicking through I striaght away see examples of this in config.py (missing module doc string), helpers.py (same), libraries.py (has incomplete method doc strings). This is not an exhaustive list but just ones I saw when going through the source code.

Part of JOSS review: https://github.com/openjournals/joss-reviews/issues/8814#issuecomment-3451697350
