### 2025-12-04 #8814 @WillJRoper
I have now reviewed the codebase along with a read of the docs and testing the installation, examples work as expected. There are some minor documentation related issues to be addressed and the contributing documentation needs to be improved to suifficiently tick off that check box (IMO).

The test coverage of 70% could be increased but gives me reasonable confidence the source code is functionally sound enough to do what the package promises without major issue.

There are parts of the code I (personally) would have approached differently but they functionally perform the promised job, and I don't think it is within scope of my review to comment line by line on the source. My main comment would have been about the namespace pollution which has now been addressed after @christopherlovell's comment and is a vast improvement on the previous approach to unit packages.

On the paper front, the artefact linked above has a missing full stop between lines 37-38, if this hasn't already been fixed it should be. I also agree with @christopherlovell that some headings or stucture improvements could be made. The paper does bleed from a statement of need to a description of the packages content and then into a lot of detail about the LICK/LIS systems. From "Too address" can be one section and then you can also highlight the related software section (which could be expanded to include other similar or related packages, e.g. `sedpy` or the DESI equivalent of the STSCI codes).

With these issues addressed and the issues on the repo addressed and closed I would be happy signing off on `pyphot`.
