### 2025-12-05 27ea972
Documentation update for contribution guidelines
