+++ CONTRIBUTING.md
@@ -1,1 +1,3 @@
 Please open a new issue or new pull request for bugs, feedback, or new features you would like to see. If there is an issue you would like to work on, please leave a comment, and we will be happy to assist. New contributions and contributors are very welcome!
+
+If you want to contribute, please look at our [contributing guide](https://mfouesneau.github.io/pyphot/contributing) and our [code of conduct](https://github.com/mfouesneau/pyphot/blob/master/CODE_OF_CONDUCT.md).
