### 2025-12-05 82f5a89
removed redundant try clause
