+++ pyphot/io/fits.py
@@ -5,118 +5,115 @@
 
 """
 
-from typing import Any, Tuple, Union, Optional
-from astropy.io import fits
-from astropy.io.fits import Header
-from astropy.io.fits import TableHDU, BinTableHDU
+from typing import Any, Optional, Tuple, Union
+
 import numpy.typing as npt
 import pandas as pd
+from astropy.io import fits
+from astropy.io.fits import BinTableHDU, Header, TableHDU
+
 from .header import HeaderInfo
 
 
-try:
-    from astropy.io import fits
+def fix_endian_issue(arr: Union[npt.NDArray, Any]) -> npt.NDArray:
+    """Fix endian issue in array which happens often when reading FITS files"""
+    return arr.astype(arr.dtype.newbyteorder())
 
-    def fix_endian_issue(arr: Union[npt.NDArray, Any]) -> npt.NDArray:
-        """Fix endian issue in array which happens often when reading FITS files"""
-        return arr.astype(arr.dtype.newbyteorder())
 
-    def fits_read_header(hdr: Header) -> HeaderInfo:
-        """
-        Convert pyfits header into dictionary with relevant values
+def fits_read_header(hdr: Header) -> HeaderInfo:
+    """
+    Convert pyfits header into dictionary with relevant values
 
-        Parameters
-        ----------
+    Parameters
+    ----------
 
-        hdr: pyftis.Header
-            fits unit
+    hdr: pyftis.Header
+        fits unit
 
-        Returns
-        -------
-        headerinfo: HeaderInfo
-            extracted information from header
-        """
-        header = {}
-        alias = {}
-        units = {}
-        comments = {}
+    Returns
+    -------
+    headerinfo: HeaderInfo
+        extracted information from header
+    """
+    header = {}
+    alias = {}
+    units = {}
+    comments = {}
 
-        # generic cards
-        genTerms = [
-            "XTENSION",
-            "BITPIX",
-            "NAXIS",
-            "NAXIS1",
-            "NAXIS2",
-            "PCOUNT",
-            "GCOUNT",
-            "TFIELDS",
-            "EXTNAME",
-        ]
-        fieldTerms = ["TTYPE", "TFORM", "TUNIT", "ALIAS"]
+    # generic cards
+    genTerms = [
+        "XTENSION",
+        "BITPIX",
+        "NAXIS",
+        "NAXIS1",
+        "NAXIS2",
+        "PCOUNT",
+        "GCOUNT",
+        "TFIELDS",
+        "EXTNAME",
+    ]
+    fieldTerms = ["TTYPE", "TFORM", "TUNIT", "ALIAS"]
 
-        for card in hdr.cards["TTYPE*"]:
-            name = card.value
-            comments[name] = card.comment
-            u = hdr.get(card.keyword.replace("TYPE", "UNIT"), None)
-            if u is not None:
-                units[name] = u
+    for card in hdr.cards["TTYPE*"]:
+        name = card.value
+        comments[name] = card.comment
+        u = hdr.get(card.keyword.replace("TYPE", "UNIT"), None)
+        if u is not None:
+            units[name] = u
 
-        # for k, val, _ in hdr.ascard['ALIAS*']:
-        for card in hdr.cards["ALIAS*"]:
-            k = card.keyword
-            val = card.value
-            al, orig = val.split("=")
-            alias[al] = orig
+    # for k, val, _ in hdr.ascard['ALIAS*']:
+    for card in hdr.cards["ALIAS*"]:
+        k = card.keyword
+        val = card.value
+        al, orig = val.split("=")
+        alias[al] = orig
 
-        # other specific keywords: COMMENT, HISTORY
-        header_comments = []
-        header_history = []
-        for k, v in hdr.items():
[…]
