### 2025-12-05 02ebb05
fix missing module docstring
