+++ _sources/pyphot.unit_adapters.rst
@@ -1,0 +1,74 @@
+pyphot.unit\_adapters package
+=============================
+
+.. automodule:: pyphot.unit_adapters
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   pyphot.unit_adapters.ezunits
+
+Submodules
+----------
+
+pyphot.unit\_adapters.astropy\_adapter module
+---------------------------------------------
+
+.. automodule:: pyphot.unit_adapters.astropy_adapter
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.astropy\_units module
+-------------------------------------------
+
+.. automodule:: pyphot.unit_adapters.astropy_units
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.ezunits\_adapter module
+---------------------------------------------
+
+.. automodule:: pyphot.unit_adapters.ezunits_adapter
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.pint\_adapter module
+------------------------------------------
+
+.. automodule:: pyphot.unit_adapters.pint_adapter
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.pint\_units module
+----------------------------------------
+
+.. automodule:: pyphot.unit_adapters.pint_units
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.typing module
+-----------------------------------
+
+.. automodule:: pyphot.unit_adapters.typing
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.unit\_adapters.units\_adapter module
+-------------------------------------------
+
+.. automodule:: pyphot.unit_adapters.units_adapter
+   :members:
+   :show-inheritance:
+   :undoc-members:
