### 2025-12-10 a291bf4
Enhance library initialization with glob pattern
Added glob_pattern parameter to the constructor and updated directory filtering logic.
