### 2025-12-10 a98f14f
Fix formatting of glob_pattern parameter in init
