### 2025-12-10 84f2a9d
Fix assignment of glob_pattern in constructor
