### 2025-12-10 5477572
Fix glob pattern handling in libraries.py
