### 2026-01-15 ed16949
migrating to h5py & adding py314 test env
