### 2026-01-15 f2db310
include h5py as default for hdf backend lib.
