+++ pyphot/io/hdf_h5py.py
@@ -5,13 +5,15 @@
 
 """
 
-from typing import Optional, Literal, Union
-import pandas as pd
 from os import PathLike
+from os.path import basename
+from typing import Literal, Optional, Union
+
+import h5py
+import numpy as np
+import pandas as pd
 
 from .header import HeaderInfo
-
-import tables
 
 
 def _decode_string_ifneeded(s: str) -> str:
@@ -27,7 +29,7 @@ def from_hdf5(
     *,
     silent: bool = True,
     **kwargs,
-):
+) -> tuple[pd.DataFrame, HeaderInfo]:
     """Generate the corresponding ascii Header that contains all necessary info
 
     Parameters
@@ -46,13 +48,22 @@ def from_hdf5(
     hdr: str
         string that will be be written at the beginning of the file
     """
-    with tables.open_file(filename, **kwargs) as source:
+    with h5py.File(filename, **kwargs) as source:
         tablename = tablename or "/"
         if not tablename.startswith("/"):
             tablename = "/" + tablename
 
-        node = source.get_node(tablename)
-        attrs = node._v_attrs
+        node = source[tablename]
+        if not node:
+            raise ValueError(
+                f"Table '{tablename}' not found in file '{filename}'"
+            )
+        if not isinstance(node, h5py.Dataset):
+            raise TypeError(
+                f"Node '{tablename}' is not a dataset (found {type(node)})"
+            )
+
+        attrs = node.attrs
         if not silent:
             print(f"\tLoading table: {tablename}")
 
@@ -61,35 +72,36 @@ def from_hdf5(
 
         # read header
         exclude = ["NROWS", "VERSION", "CLASS", "EXTNAME", "TITLE"]
-        for k in attrs._v_attrnames:
+        for k, v in attrs.items():
             if k not in exclude:
                 if not k.startswith("FIELD") and not k.startswith("ALIAS"):
-                    header[k] = _decode_string_ifneeded(attrs[k])
+                    header[k] = _decode_string_ifneeded(v)
                 elif k.startswith("ALIAS"):
-                    c0, c1 = _decode_string_ifneeded(attrs[k]).split("=")
+                    c0, c1 = _decode_string_ifneeded(v).split("=")
                     aliases[c0] = c1
 
-        if attrs["TITLE"] not in ["", "None", "Noname", None]:
-            header["NAME"] = attrs["TITLE"]
+        title = attrs.get("TITLE", "")
+        if title not in ["", "None", "Noname", None]:
+            header["NAME"] = _decode_string_ifneeded(title)
         else:
-            header["NAME"] = f"{filename:s}/{node._v_name:s}"
+            header["NAME"] = f"{filename:s}/{node.name:s}"
 
         # read column meta
         units = {}
         desc = {}
 
-        colnames = node.colnames  # type: ignore / it does exist
+        colnames = node.dtype.names
         for k, colname in enumerate(colnames):
-            _u = getattr(attrs, f"FIELD_{k:d}_UNIT", None)
-            _u = getattr(attrs, f"{colname:s}_UNIT", _u)
-            _d = getattr(attrs, f"FIELD_{k:d}_DESC", None)
-            _d = getattr(attrs, f"{colname:s}_DESC", _d)
+            _u = attrs.get(f"FIELD_{k:d}_UNIT", None)
+            _u = attrs.get(f"{colname:s}_UNIT", _u)
+            _d = attrs.get(f"FIELD_{k:d}_DESC", None)
+            _d = attrs.get(f"{colname:s}_DESC", _d)
             if _u is not None:
                 units[colname] = _decode_string_ifneeded(_u)
             if _d is not None:
[…]
