+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyphot"
-version = "2.0.0"
+version = "2.1.0"
 description = "A tool for computing photometry from spectra"
 readme = "README.rst"
 requires-python = ">=3.9"
