### 2026-01-15 3072aff
Migrate to h5py (#94)
* migrating to h5py & adding py314 test env

* include h5py as default for hdf backend lib.

* Add python 3.14 label

* version update and changelog
