### 2026-01-15 53fef3f
fix pixi task defintion
