+++ pyproject.toml
@@ -76,7 +76,7 @@ docs = [
   "matplotlib",
   "numpydoc"]
 
-[tool.pixi.project]
+[tool.pixi.workspace]
 channels = ["conda-forge"]
 platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
 
