### 2026-01-15 5a7cad7
structure update integrating review feedback
