+++ paper/paper.bib
@@ -152,3 +152,70 @@ archivePrefix = {arXiv},
 }
 
 
+@ARTICLE{Korn2005,
+       author = {{Korn}, A.~J. and {Maraston}, C. and {Thomas}, D.},
+        title = "{The sensitivity of Lick indices to abundance variations}",
+      journal = {\aap},
+     keywords = {radiative transfer, line: formation, stars: late-type, galaxies: abundances, galaxies: evolution, galaxies: fundamental parameters, Astrophysics},
+         year = 2005,
+        month = aug,
+       volume = {438},
+       number = {2},
+        pages = {685-704},
+          doi = {10.1051/0004-6361:20042126},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0504574},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2005A&A...438..685K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+@ARTICLE{Worthey1997,
+       author = {{Worthey}, Guy and {Ottaviani}, D.~L.},
+        title = "{H{\ensuremath{\gamma}} and H{\ensuremath{\delta}} Absorption Features in Stars and Stellar Populations}",
+      journal = {\apjs},
+     keywords = {GALAXIES: EVOLUTION, GALAXIES: STELLAR CONTENT, LINE: PROFILES, STARS: ABUNDANCES, Galaxies: Evolution, Galaxies: Stellar Content, Line: Profiles, Stars: Abundances},
+         year = 1997,
+        month = aug,
+       volume = {111},
+       number = {2},
+        pages = {377-386},
+          doi = {10.1086/313021},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/1997ApJS..111..377W},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Thomas2011,
+       author = {{Thomas}, Daniel and {Maraston}, Claudia and {Johansson}, Jonas},
+        title = "{Flux-calibrated stellar population models of Lick absorption-line indices with variable element abundance ratios}",
+      journal = {\mnras},
+     keywords = {stars: evolution, Galaxy: abundances, globular clusters: general, galaxies: abundances, galaxies: star clusters: general, galaxies: stellar content, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2011,
+        month = apr,
+       volume = {412},
+       number = {4},
+        pages = {2183-2198},
+          doi = {10.1111/j.1365-2966.2010.18049.x},
+archivePrefix = {arXiv},
+       eprint = {1010.4569},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011MNRAS.412.2183T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Spiniello2014,
+       author = {{Spiniello}, Chiara and {Trager}, Scott and {Koopmans}, L{\'e}on V.~E. and {Conroy}, Charlie},
+        title = "{The stellar IMF in early-type galaxies from a non-degenerate set of optical line indices}",
+      journal = {\mnras},
+     keywords = {gravitational lensing: strong, galaxies: elliptical and lenticular, cD, galaxies: evolution, galaxies: kinematics and dynamics, galaxies: structure, dark matter, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2014,
+        month = feb,
+       volume = {438},
+       number = {2},
[…]
