+++ paper/paper.md
@@ -7,7 +7,7 @@ tags:
 authors:
   - name: Morgan Fouesneau
     orcid: 0000-0001-9256-5516
-    affiliation: 1 
+    affiliation: 1
     corresponding: true # (This is how to denote the corresponding author)
   - name: Ariane Lançon
     affiliation: 2
@@ -31,17 +31,22 @@ Synthetic photometry is a crucial task in astronomy, used to estimate the flux a
 
 Photometry involves measuring the flux or intensity of light emitted by astronomical objects. It is a fundamental technique used in astronomy to study any object that emits light, such as stars, galaxies, and quasars. Photometry provides information about the brightness, color, and spectral energy distribution of these objects, which is essential for understanding their physical properties and evolutionary history.
 
-The type of detector and the units of flux can vary depending on the wavelength range being considered. One of the challenges in computing photometry is ensuring that one performs the calculations in the correct units and handling conversions between different units (e.g., Vega to AB magnitude or vice-versa). The online documentation[^3] of `pyphot` provides the mathematical details and references
+The type of detector and the units of flux can vary depending on the wavelength range being considered. One of the challenges in computing photometry is ensuring that one performs the calculations in the correct units and handling conversions between different units (e.g., Vega to AB magnitude or vice-versa).
 
-To address this challenge, the `pyphot` tool provides a comprehensive set of functions and utilities for computing photometry from spectra. It includes built-in support for unit handling and conversions based on either the implementations of physical units in the `Astropy` package [@astropy] (`astropy.units`) or in the `Pint` package[^2]. It also provides a database of pre-defined filter transmission curves, lick index definitions [@worthey1994], and interfaces with the [SVO filter profile service](http://svo2.cab.inta-csic.es/theory/fps/) [@svofps2020] to enable a broader range of filters.
+# The pyphot package
 
-In addition to calculating photometry, `pyphot` provides a set of tools to define and calculate lick indices. The Lick system of spectral line indices is one of the most commonly used methods for estimating ages and metallicities of unresolved (integrated light) stellar populations. The calibration of the Lick system is complicated because the original Lick spectra were not flux calibrated, so there are usually systematic effects due to differences in continuum shape. Proper calibration involves observing many original lick standard stars and deriving offsets to the standard system.
+To address this challenge, the `pyphot` tool provides a comprehensive set of functions and utilities for computing photometry from spectra. The online documentation[^3] of `pyphot` provides the mathematical details and references. It includes built-in support for unit handling and conversions based on either the implementations of physical units in the `Astropy` package [@astropy] (`astropy.units`) or in the `Pint` package[^2]. Users can define their transmission curves but we also provide a database of pre-defined filter transmission curves[^4], and interfaces with the [SVO filter profile service](http://svo2.cab.inta-csic.es/theory/fps/) [@svofps2020] to enable a broader range of publicly available photometric filters.
 
[…]
