+++ CITATION.cff
@@ -5,6 +5,6 @@ authors:
   given-names: "Morgan"
   orcid: "https://orcid.org/0000-0001-9256-5516"
 title: "pyphot"
-version: 2.0.0
-date-released: 2025-11-15
+version: 2.1.0
+date-released: 2026-01-15
 url: "https://github.com/mfouesneau/pyphot"
