+++ _sources/modules.rst
@@ -1,0 +1,7 @@
+API
+===
+
+.. toctree::
+   :maxdepth: 4
+
+   pyphot
