+++ _sources/pyphot.io.rst
@@ -1,0 +1,58 @@
+pyphot.io package
+=================
+
+.. automodule:: pyphot.io
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+Submodules
+----------
+
+pyphot.io.ascii module
+----------------------
+
+.. automodule:: pyphot.io.ascii
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.fits module
+---------------------
+
+.. automodule:: pyphot.io.fits
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.hdf\_h5py module
+--------------------------
+
+.. automodule:: pyphot.io.hdf_h5py
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.hdf\_pytables module
+------------------------------
+
+.. automodule:: pyphot.io.hdf_pytables
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.header module
+-----------------------
+
+.. automodule:: pyphot.io.header
+   :members:
+   :show-inheritance:
+   :undoc-members:
+
+pyphot.io.votable module
+------------------------
+
+.. automodule:: pyphot.io.votable
+   :members:
+   :show-inheritance:
+   :undoc-members:

[…]
