### 2026-01-29 #8814 @JBorrow
Looks like this is ready for post-reviewer checks; thanks so much @christopherlovell and @WillJRoper for volunteering and taking the time to review, your contributions are really what enables JOSS to function!
