### 2026-01-29 #8814 @JBorrow
@mfouesneau, there are some 'Additional author tasks' above. Please let me know when those are completed; for now I will begin copy-editing the submission.
