### 2026-01-29 #8814 @JBorrow
Copyediting comments (line numbers are baked into the PDF above):

- `pyphot` or 'Pyphot': you use these interchangeably throughout the paper, but when referring to the package by name consistently using 'Pyphot' would be better. You can choose to use the `pyphot` style throughout if you wish; just do so consistently.
- A similar comment about the capitalization of `Astropy` and `Pint`; there's no need for these to be monospaced when using them as names in the text. `astropy.units` makes sense, as this is referring to a specific package/import inside the Astropy suite.
- On line 54, there is a double lcosing bracket that has no opening companion.
- At the end of line 65 there is a `pyphot` on a line by itself.
