### 2026-03-02 #8814 @JBorrow
Great, thank you. There are a couple extra tasks above ('Additional Author Tasks After Review is Complete'), once those are complete I can pass this to the EiC.
