### 2026-03-02 #8814 @mfouesneau
- added ORCID second author to the paper.
      - released v2.1.0
      - with DOI https://doi.org/10.5281/zenodo.18256672
