### 2026-03-02 #8814 @JBorrow
@mfouesneau the archive currently only lists you (not including Ariane Lancon) as an author, and the title should match the paper exactly (i.e. it should be pyphot: A tool for computing photometry from spectra).
