### 2026-03-03 #8814 @mfouesneau
@JBorrow, I see, sorry I missed this. It is now fixed in the citation.cff file:
- author list
- title
version 2.1.1  released and archived
