### 2026-03-06 #8814 @warrickball
Thanks once again to @christopherlovell and @WillJRoper for reviewing and @JBorrow for editing this submission. JOSS simply wouldn't be possible without its community of volunteers.
